### 2018-08-02 #866 @whedon
[PRE REVIEW]: Comet Time Series (CometTS) Visualizer
**Submitting author:** @jshermeyer (<a href="http://orcid.org/0000-0002-8143-2790">Jacob Shermeyer</a>)
**Repository:** <a href="https://github.com/CosmiQ/CometTS" target ="_blank">https://github.com/CosmiQ/CometTS</a>
**Version:** v1.0
