### 2018-08-17 #866 @Kevin-Mattheus-Moerman
@jshermeyer before we proceed, can you help address the issue with you paper. It seems it cannot find the references to `Holden:2015`. Can you check this?
