### 2018-09-10 #866 @Kevin-Mattheus-Moerman
:point_up: @jshermeyer can you check that paper reference to help compile your paper?
