### 2018-09-12 #866 @jshermeyer
Apologies, have been traveling. Will address this next week. Thanks for the reminder!

-Jake

> On Sep 10, 2018, at 12:16 PM, Kevin Mattheus Moerman <notifications@github.com> wrote:
>
> ☝️ @jshermeyer can you check that paper reference to help compile your paper?
>
