### 2018-09-19 f213c22
Update paper.md
