### 2018-09-19 #866 @Kevin-Mattheus-Moerman
@jshermeyer you can also do '@whedon generate pdf'
