### 2018-09-19 6b67c49
Update paper.md
