### 2018-09-19 6f0e31f
Update paper.md
