### 2018-09-19 c421d4d
Update paper.md
