### 2018-09-19 #866 @jshermeyer
@arfon @Kevin-Mattheus-Moerman Thanks both, proof looks good to me.
