### 2018-10-01 #866 @jshermeyer
@Kevin-Mattheus-Moerman Is there a listing of persons that you would recommend? Or is this just anyone from an academic/research community?
