### 2018-10-10 #866 @Kevin-Mattheus-Moerman
@jshermeyer we can invite anybody from the academic research community.
