### 2018-10-10 #866 @Kevin-Mattheus-Moerman
@campagnola would you be interested in reviewing this submission for [JOSS](http://joss.theoj.org/)?
