### 2018-10-24 #866 @Kevin-Mattheus-Moerman
@zhampel @jgrss @slandersson @mmann1123 @avanetten would you be interested in reviewing this submission for the [Journal of Open Source Software (JOSS)](http://joss.theoj.org/)? The [review process](https://joss.readthedocs.io/en/latest/review_criteria.html) (which takes place on GitHub) focuses largely on [the software](https://github.com/CosmiQ/CometTS) and a [short paper](https://github.com/openjournals/joss-papers/blob/joss.00866/joss.00866/10.21105.joss.00866.pdf).
Let me know if you have questions.
