### 2018-10-24 #866 @zhampel
@Kevin-Mattheus-Moerman Yes, I'd be happy to review the CometTS submission.
