### 2018-10-24 #866 @Kevin-Mattheus-Moerman
To avoid further delay in processing this submission I'll now start review. I do welcome additional reviewers if any are found, I'll add them on the review issue itself. Thanks all. Let's get started.
