### 2018-10-24 #1047 @whedon
[REVIEW]: Comet Time Series (CometTS) Visualizer
**Submitting author:** @jshermeyer (<a href="http://orcid.org/0000-0002-8143-2790">Jacob Shermeyer</a>)
**Repository:** <a href="https://github.com/CosmiQ/CometTS" target ="_blank">https://github.com/CosmiQ/CometTS</a>
**Version:** <v1.2.0>

@zhampel, and @rmsare please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:

- [x] **Version:** <v1.2.0>

- [x] **Version:** <v1.2.0>

- [x] **Version:** <v1.2.0>
