### 2018-10-30 #866 @dmittman
Fascinating.
