### 2018-10-30 #1047 @Kevin-Mattheus-Moerman
@rhiever, @zhampel, @dmittman thanks again for acting as reviewers. You all have a set of tick boxes at the top of this issue which will guide you through the review process (which you can tick if you've accepted [this invitation](https://github.com/openjournals/joss-reviews/invitations)). Please let me know if you have any questions. :rocket: :robot:
