### 2018-10-30 #866 @arfon
> In any event, this was the pre-review issue to find reviewers, @Kevin-Mattheus-Moerman is the editor, and he should add you to the actual review issue: #1047

Closing this as the review is moving forward in #1047
