### 2018-10-31 f1ff3b5
Create CONTRIBUTING.MD
