+++ CONTRIBUTING.MD
@@ -1,0 +1,21 @@
+ ## How to contribute to CometTS
+
+#### **Did you find a bug?**
+
+* **Ensure the bug was not already reported** by searching on GitHub under [Issues](https://github.com/CosmiQ/CometTS/issues).
+
+* If you're unable to find an open issue addressing the problem, [open a new one](https://github.com/CosmiQ/CometTS/issues/new). Be sure to include a **title and clear description**, as much relevant information as possible, and a **code sample** or an **executable test case** demonstrating the expected behavior that is not occurring.
+
+#### **Did you write a patch that fixes a bug?**
+
+* Open a new GitHub pull request with the patch.
+
+* Ensure the description clearly describes the problem and solution. Include the relevant issue number if applicable.
+
+#### **Do you intend to add a new feature or change an existing one?**
+
+* Suggest your change in [Issues](https://github.com/CosmiQ/CometTS/issues) section.
+
+#### **Do you have questions about the source code?**
+
+* Ask any question in the [Issues](https://github.com/CosmiQ/CometTS/issues) section.
