### 2018-10-31 #1047 @jshermeyer
@Kevin-Mattheus-Moerman
Added a CONTRIBUTION.MD.  Code of conduct seems overkill, if I need to post a code of conduct calling for basic civility in my little codebase I think the world is doomed.

@zhampel
Thanks for the notes.  Will make some edits when another review comes in.

_I think it would be better to intro something like ‘user defined polygon identifying a region of interest (ROI)’ and use ROI for the remainder of the paper._

Agreed, will make this change then when other reviews come back.

_Q: Any limits to ‘user defined area of interest’?_

Just computational strength, the larger the polygon the more slowly the processing.  I believe CometTS will not like polygons that are larger than the imagery you provide, but it can tolerate and ignore nodata/ blank space/ masked areas.

_Figures: please provide descriptive captions to the three figures._
Good idea.
