### 2018-11-02 bf90587
Functions for autoregressive integrated moving average analysis
https://arxiv.org/abs/1807.05854
