+++ ARIMA.py
@@ -1,0 +1,197 @@
+import sys
+import os
+import glob
+import gdal
+import rasterstats
+import rasterio
+from rasterstats import zonal_stats
+import numpy as np
+import pandas as pd
+import geopandas as gpd
+import fiona
+from shapely.geometry import shape
+import shapely.wkt
+import string
+import matplotlib.pyplot as plt
+import seaborn as sns; sns.set(color_codes=True)
+import datetime
+from ipywidgets import interact, interactive, fixed, interact_manual
+from IPython.display import Javascript, display, clear_output
+import ipywidgets as widgets
+from tqdm import tqdm_notebook as tqdm
+import matplotlib.dates as mdates
+from scipy.signal import gaussian
+from scipy.ndimage import filters
+from affine import Affine
+from fnmatch import fnmatch
+
+
+### Functions for an auto-regressive integrated moving average analysis using CometTS on a time series of satellite imagery.
+
+
+def run_plot_TS(main_gdf, figsize = (12,6), y_val_alpha = 1, scatter_alpha = 1, 
+             error_alpha = 0.2, y_label = "Brightness", x_label = "Date", title_label= "Brightness over time - Census Tract 9509, Yabucoa Municipio" 
+             , outname="/Outname.csv",figname = '', custom_x_axis = True, show_grid = True, show_legend = True, min_count=0.5,ymin=0,ymax=5000):
+    print "Plotting..." 
+    C=0
+    for item in main_gdf['ID'].unique():
+        C+=1
+        plt.style.use('fivethirtyeight')
+        fig, ax = plt.subplots(figsize=figsize)
+        gdf3=main_gdf[(main_gdf.ID == item)]
+        title= title_label
+        gdf3=gdf3.sort_values(['date'])
+        gdf3=TS_Trend(gdf3,CMA_Val=5,CutoffDate="2017/08/31")
+        gdf3=gdf3.sort_values(['date'])
+        x = gdf3['date']
+        xdate= x.astype('O')
+        xdate=mdates.date2num(xdate)
+        y  = gdf3['mean']
+        if 'SeasonalForecast' in gdf3.columns:
+            if C==1:
+                gdf_holder=gdf3
+            else:
+                gdf_holder=gdf_holder.append(gdf3)
+            T= gdf3['SeasonalForecast']
+            if 'observations' in gdf3.columns:
+                count= gdf3['observations']
+            err_plus = gdf3['mean'] + gdf3['std']
+            err_minus = gdf3['mean'] - gdf3['std']
+            #Set the min_count value as a value from 0 to 1 (0 to 100%)
+            #ax.set_ylim([ymin,ymax])
+
+            #This will filter out observations where over n% of a polygon is masked out
+            if min_count > 0:
+                z = gdf3['count']
+                zmax=gdf3['count'].max()
+                z=z/zmax
+                xdate = xdate[z >= min_count]
+                y = y[z >= min_count]
+                if 'observations' in gdf3.columns:
+                    count=count[z>= min_count]
+                err_plus = err_plus[z >= min_count]
+                err_minus = err_minus[z >= min_count]
+                
+                
+            if len(xdate)==len(gdf3['Trend']):
+                # plot regression line, if desired
+                idx = np.isfinite(xdate) & np.isfinite(y)
+                p2 = np.poly1d(np.polyfit(xdate[idx], y[idx], 1))
+                ax.plot(xdate, gdf3['Trend'], '-',label="Linear Forecast", color='#00C5B0', alpha=y_val_alpha)
+
[…]
