### 2018-11-21 #1047 @Kevin-Mattheus-Moerman
@dmittman Thanks for your reply. I understand. Thanks for taking the time to check out this project, and for recommending another reviewer.
@lewismc would you be interested in reviewing this work for [JOSS](http://joss.theoj.org/)?
