### 2018-11-30 #1047 @Kevin-Mattheus-Moerman
@lewismc the review focuses on [the software](https://github.com/CosmiQ/CometTS) and [the short paper](https://github.com/openjournals/joss-papers/blob/joss.01047/joss.01047/10.21105.joss.01047.pdf). The instructions and tick boxes  (see the top of this issue) will guide you through the process. Let me know if you still have questions. You can also consult our review criteria here: https://joss.readthedocs.io/en/latest/review_criteria.html.
