### 2018-12-14 #1047 @zhampel
@jshermeyer here are my comments regarding the structure of CometTS's code. I found your instructions easy to follow with the tools you have built. Your code does work as advertised, though I do think there are a changes that should be made to the formatting of the code, including commenting and other coding standards that should be upheld. Furthermore, I think your work could be made much stronger as an extensible project, by migrating the functionality of your python notebooks to a more API-like structure.

Code Review
--------------

General
---------

- MacOS fails to pip install due to gdal. Needs `brew install gdal` prior to `pip install -r Requirements.txt`. Works on my MacOS Mojave laptop.

- DownThemAll add-on link points to a non-existent page.

- Ensuring a proper directory structure is clear for the example you gave, but is that the case with other datasets? Not being a geospatial data expert, is there a resource that provides guidance on proper structuring, or perhaps a universal industry standard?

- Had difficulty with GDAL install (who doesn’t?). Perhaps best to include gdal and subsequent deps on gdal at the end of the requirements list, so that all other packages install first. This may assist in debugging install.

- Also recommend including link to building GDAL with python-bindings in Installation section. Had success using following: http://trac.osgeo.org/gdal/wiki/BuildingOnUnix

- Please include a python style .gitignore file that excludes the CometTS python notebooks.

Code Specifics
----------------

- Comment functions in code:

	- Recommend reformatting for python3 compatibility

	- Recommended to import functions explicitly, instead of * for ease of tracking

	- Header description for each fn

	- Inputs and outputs description

	- All input argument declarations should be initialized to None, ‘’, etc

	- Run flake8 to format code properly, will make it easier to read too ;)

	- Multi_studyAreas_FullStats.csv is in same dir as notebooks. Plot_Results.ipynb looks elsewhere for it. Consider moving it into the example data dir. Yet, when trying to use this example csv data, I get an error that there is no key `count`. Did work on my generated csv data though.

	- For plotting, should have more scalable y-axis, as I chose a rather small city (Santa Fe, NM) to test

	- Difficult to debug errors, esp wrt formatting data issues

	- What are the temp*.vrt files? Are they part of the project?
