+++ py_temp.py
@@ -1,0 +1,524 @@
+# Module for plotting results only!
+
+from CometTS import *
+import CometTS  # See ComeTS.py to edit functions
+from affine import Affine
+from scipy.ndimage import filters
+from scipy.signal import gaussian
+import matplotlib.dates as mdates
+from tqdm import tqdm_notebook as tqdm
+import ipywidgets as widgets
+from IPython.display import Javascript, display, clear_output
+from ipywidgets import interact, interactive, fixed, interact_manual
+import datetime
+import sys
+import os
+import glob
+import gdal
+import rasterstats
+import rasterio
+from rasterstats import zonal_stats
+import numpy as np
+import pandas as pd
+import geopandas as gpd
+import fiona
+from shapely.geometry import shape
+import shapely.wkt
+import string
+import matplotlib.pyplot as plt
+import seaborn as sns
+sns.set(color_codes=True)
+
+# Run plotting from CSV output ONLY
+
+input_CSV = "/AOI_InterestingPlaces_SWIR2/NO_InterestingPlaces_UTM_FullStats.csv"
+input_CSV2 = "/AOI_Interesting_Places_Green/NO_InterestingPlaces_UTM_FullStats.csv"
+input_CSV3 = "/AOI_InterestingPlaces_NIR/NO_InterestingPlaces_UTM_FullStats.csv"
+%matplotlib inline
+
+
+def gen_dual_plot(
+        input_csv,
+        input_csv2,
+        title_label="Spectral Response Over Time - ID: "):
+    CSVs = [input_csv, input_csv2]
+    gdfs = ["gdf_1", "gdf_2"]
+    count = 0
+    for csv in CSVs:
+        df = pd.read_csv(csv)
+        df = df.sort_values(['date'])
+        geometry = df['geometry'].map(shapely.wkt.loads)
+        crs = {'init': 'epsg:2263'}
+        gdfs[count] = gpd.GeoDataFrame(df, crs=crs, geometry=geometry)
+        gdfs[count]['date'] = pd.to_datetime(
+            gdfs[count]['date'], infer_datetime_format=True)
+        count += 1
+    run_dual_plot(
+        gdfs[0],
+        gdfs[1],
+        title_label=title_label,
+        show_grid=False,
+        ymax=3000,
+        y_label="Surface Reflectance x 10000")
+
+
+def gen_tri_plot(
+        input_csv,
+        input_csv2,
+        input_csv3,
+        title_label="Spectral Response Over Time - ID: "):
+    CSVs = [input_csv, input_csv2, input_csv3]
+    gdfs = ["gdf_1", "gdf_2", "gdf_3"]
+    count = 0
+    for csv in CSVs:
+        df = pd.read_csv(csv)
+        df = df.sort_values(['date'])
+        geometry = df['geometry'].map(shapely.wkt.loads)
+        crs = {'init': 'epsg:2263'}
+        gdfs[count] = gpd.GeoDataFrame(df, crs=crs, geometry=geometry)
+        gdfs[count]['date'] = pd.to_datetime(
+            gdfs[count]['date'], infer_datetime_format=True)
+        count += 1
+    run_tri_plot(
+        gdfs[0],
+        gdfs[1],
+        gdfs[2],
+        title_label=title_label,
+        show_grid=False,
+        ymax=5000,
+        y_label="Surface Reflectance x 10000")
+
+
+# def gen_plots(input_csv, title_label= "Spectral Response Over Time - ID: "):
+#    df = pd.read_csv(input_csv)
+#    df=df.sort_values(['date'])
+#    geometry = df['geometry'].map(shapely.wkt.loads)
+#    crs = {'init': 'epsg:2263'}
[…]
