+++ ARIMA.py
@@ -1,20 +1,22 @@
 import matplotlib.dates as mdates
-from tqdm import tqdm_notebook as tqdm
+from tqdm import tqdm as tqdm
 import datetime
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
 import seaborn as sns
 from CometTS import interpolate_gaps
+import argparse
+import os
 sns.set(color_codes=True)
 
 
-# Functions for an auto-regressive integrated moving average analysis
+# Functions for a seasonal auto-regressive integrated moving average analysis
 # using CometTS on a time series of satellite imagery.
 
 
 def run_plot_TS(
-        main_gdf,
+        ARIMA_GDF,
         figsize=(
             12,
             6),
@@ -23,8 +25,7 @@ def run_plot_TS(
     error_alpha=0.2,
     y_label="Brightness",
     x_label="Date",
-    title_label="Brightness over time - Census Tract 9509, Yabucoa Municipio",
-    outname="/Outname.csv",
+    title_label="ARIMA_Trend",
     figname='',
     custom_x_axis=True,
     show_grid=True,
@@ -34,19 +35,21 @@ def run_plot_TS(
         ymax=5000):
     print("Plotting...")
     C = 0
-    for item in main_gdf['ID'].unique():
+    # ARIMA GDF is the output GDF from TS_Trend function
+    for item in ARIMA_GDF['ID'].unique():
         C += 1
         plt.style.use('fivethirtyeight')
         fig, ax = plt.subplots(figsize=figsize)
-        gdf3 = main_gdf[(main_gdf.ID == item)]
+        gdf3 = ARIMA_GDF[(ARIMA_GDF.ID == item)]
         title = title_label
-        gdf3 = gdf3.sort_values(['date'])
-        gdf3 = TS_Trend(gdf3, CMA_Val=5, CutoffDate="2017/08/31")
+        # gdf3 = gdf3.sort_values(['date'])
+        # gdf3 = TS_Trend(gdf3, CMA_Val=5, CutoffDate="2017/08/31")
         gdf3 = gdf3.sort_values(['date'])
         x = gdf3['date']
         xdate = x.astype('O')
         xdate = mdates.date2num(xdate)
         y = gdf3['mean']
+        anomaly = gdf3['Anomaly']
         if 'SeasonalForecast' in gdf3.columns:
             if C == 1:
                 gdf_holder = gdf3
@@ -55,8 +58,8 @@ def run_plot_TS(
             T = gdf3['SeasonalForecast']
             if 'observations' in gdf3.columns:
                 count = gdf3['observations']
-            err_plus = gdf3['mean'] + gdf3['std']
-            err_minus = gdf3['mean'] - gdf3['std']
+            err_plus = gdf3['SeasonalError_Pos']
+            err_minus = gdf3['SeasonalError_Neg']
             # Set the min_count value as a value from 0 to 1 (0 to 100%)
             # ax.set_ylim([ymin,ymax])
 
@@ -107,6 +110,13 @@ def run_plot_TS(
                     s=50,
                     color='black',
                     alpha=scatter_alpha)
+                ax.scatter(
+                    xdate,
+                    anomaly,
+                    label="Anomalies",
+                    s=50,
+                    color='red',
+                    alpha=scatter_alpha)
                 #ax.scatter(xdate, y2, label="Mean", s=50, color='red',alpha=scatter_alpha,marker='x')
 
                 # if desired, plot error band
@@ -116,7 +126,7 @@ def run_plot_TS(
                     err_minus,
[…]
