+++ MANIFEST.in
@@ -1,0 +1,4 @@
+include *.txt
+include *.md
+include /VIIRS_Sample/*.tif
+include /VIIRS_Sample/*.tif
