### 2019-02-21 0ca75ac
Update MANIFEST.in
