### 2019-02-21 f86c340
Update MANIFEST.in
