### 2019-02-21 8b0e25f
Update and rename LICENSE to LICENSE.md
