+++ CometTS/ARIMA.py
@@ -1,0 +1,292 @@
+import matplotlib.dates as mdates
+from tqdm import tqdm as tqdm
+import datetime
+import numpy as np
+import pandas as pd
+import matplotlib.pyplot as plt
+import seaborn as sns
+from CometTS import interpolate_gaps
+import argparse
+import os
+sns.set(color_codes=True)
+
+
+# Functions for a seasonal auto-regressive integrated moving average analysis
+# using CometTS on a time series of satellite imagery.
+
+
+def run_plot_TS(
+        ARIMA_GDF,
+        figsize=(
+            12,
+            6),
+    y_val_alpha=1,
+    scatter_alpha=1,
+    error_alpha=0.2,
+    y_label="Brightness",
+    x_label="Date",
+    title_label="ARIMA_Trend",
+    figname='',
+    custom_x_axis=True,
+    show_grid=True,
+    show_legend=True,
+    min_count=0.5,
+    ymin=0,
+        ymax=5000):
+    print("Plotting...")
+    C = 0
+    # ARIMA GDF is the output GDF from TS_Trend function
+    for item in ARIMA_GDF['ID'].unique():
+        C += 1
+        plt.style.use('fivethirtyeight')
+        fig, ax = plt.subplots(figsize=figsize)
+        gdf3 = ARIMA_GDF[(ARIMA_GDF.ID == item)]
+        title = title_label
+        # gdf3 = gdf3.sort_values(['date'])
+        # gdf3 = TS_Trend(gdf3, CMA_Val=5, CutoffDate="2017/08/31")
+        gdf3 = gdf3.sort_values(['date'])
+        x = gdf3['date']
+        xdate = x.astype('O')
+        xdate = mdates.date2num(xdate)
+        y = gdf3['mean']
+        anomaly = gdf3['Anomaly']
+        if 'SeasonalForecast' in gdf3.columns:
+            if C == 1:
+                gdf_holder = gdf3
+            else:
+                gdf_holder = gdf_holder.append(gdf3)
+            T = gdf3['SeasonalForecast']
+            if 'observations' in gdf3.columns:
+                count = gdf3['observations']
+            err_plus = gdf3['SeasonalError_Pos']
+            err_minus = gdf3['SeasonalError_Neg']
+            # Set the min_count value as a value from 0 to 1 (0 to 100%)
+            # ax.set_ylim([ymin,ymax])
+
+            # This will filter out observations where over n% of a polygon is
+            # masked out
+            if min_count > 0:
+                z = gdf3['count']
+                zmax = gdf3['count'].max()
+                z = z / zmax
+                xdate = xdate[z >= min_count]
+                y = y[z >= min_count]
+                if 'observations' in gdf3.columns:
+                    count = count[z >= min_count]
+                err_plus = err_plus[z >= min_count]
+                err_minus = err_minus[z >= min_count]
+
+            if len(xdate) == len(gdf3['Trend']):
+                # plot regression line, if desired
+                # idx = np.isfinite(xdate) & np.isfinite(y)
+                # p2 = np.poly1d(np.polyfit(xdate[idx], y[idx], 1))
+                ax.plot(
+                    xdate,
+                    gdf3['Trend'],
+                    '-',
+                    label="Linear Forecast",
+                    color='#00C5B0',
+                    alpha=y_val_alpha)
+
+                # plot running mean regression line
+                # RM=(y.rolling(window=6,center=True, min_periods=2).median())
+                # ax.plot(xdate, RM, '-',label="Moving Median", alpha=y_val_alpha)
+
+                # Seasonal Forecast
+                T = interpolate_gaps(T, limit=3)
+                ax.plot(
+                    xdate,
[…]
