### 2019-02-21 992e4ef
Delete CometTS.py
