+++ CometTS.py
@@ -1,1039 +1,0 @@
-import os
-import glob
-import gdal
-import rasterio
-from rasterstats import zonal_stats
-import numpy as np
-import pandas as pd
-import geopandas as gpd
-import shapely.wkt
-import matplotlib.pyplot as plt
-# from tqdm import tqdm_notebook as tqdm
-import matplotlib.dates as mdates
-from scipy.signal import gaussian
-from scipy.ndimage import filters
-from fnmatch import fnmatch
-import argparse
-
-
-def Process_imagery(Path_dir, zonalpoly, NoDataValue, mask_value, maskit=True):
-    print(zonalpoly)
-    gdf = gpd.read_file(zonalpoly)
-    if maskit:
-        mask_value = str(mask_value)
-        mask_value = mask_value.split(",")
-    # Get the zonal stats
-    NoDataValue = int(NoDataValue)
-    gdf2 = Do_Zonal_Stats(Path_dir, gdf, NoDataValue, mask_value, maskit)
-    # Get number of observations
-    gdf3 = Get_Num_Obs(Path_dir, gdf, NoDataValue, mask_value, maskit)
-    for date in gdf2.iterrows():
-        if 'median' in gdf3.columns:
-            gdf2['observations'] = gdf3['median']
-
-    # Save CSV
-    print("Producing csv output...")
-    z_simple = zonalpoly.split('/')
-    z_simple = z_simple[-1].split('.')
-    z_simple = z_simple[0]
-    Path_out = os.path.dirname(os.path.abspath(Path_dir))
-
-    output = os.path.join(Path_out, z_simple + '_FullStats.csv')
-    print("CSV statistics saved here: ", output)
-    gdf2.to_csv(output)
-    for item in gdf2['ID'].unique():
-        gdf3 = gdf2[(gdf2.ID == item)]
-        output = os.path.join(
-            Path_out,
-            z_simple +
-            '_Stats_ID_' +
-            str(item) +
-            '.csv')
-        gdf3.to_csv(output)
-
-    return gdf2
-
-
-# Mask your data
-def Mask_it(rasterin, mask, geom, NoDataValue, mask_value):
-    geom = geom.bounds
-    geom = [geom[0], geom[3], geom[2], geom[1]]
-    MRO = gdal.Translate("temp1.vrt", rasterin, projWin=geom).ReadAsArray()
-    MR2 = gdal.Translate("temp2.vrt", mask, projWin=geom).ReadAsArray()
-    r = rasterio.open("temp1.vrt")
-    affineO = r.transform
-    del r
-    for item in mask_value:
-        MRO[np.where(MR2 == int(item))] = NoDataValue
-    return MRO, affineO
-
-
-def Do_Zonal_Stats(Path_dir, gdf, NoDataValue, mask_value, maskit=True):
-    data = pd.read_csv(Path_dir)
-    data = data.sort_values(['date'])
-    zonelist = []
-    print("Processing...")
-    for idx, row in tqdm(data.iterrows(), total=data.shape[0]):
-        if row['TS_Data'] == 1:
-            raster = row['File']
-            count = 0
-            date = row['date']
-            for idx, feature in gdf.iterrows():
-                count += 1
-                if maskit:
-                    MR, affine = Mask_it(
-                        raster, row['Mask'], feature['geometry'], NoDataValue, mask_value)
-                    statout = zonal_stats(
-                        feature["geometry"],
-                        MR,
-                        stats="min max median mean std percentile_25 percentile_75 count",
-                        nodata=NoDataValue,
-                        affine=affine)
-                if maskit is False:
-                    statout = zonal_stats(
-                        feature["geometry"],
-                        raster,
-                        stats="min max median mean std percentile_25 percentile_75 count",
-                        nodata=NoDataValue)
-
[…]
