### 2019-02-21 3ba931c
Delete baseeval_test.py
