### 2019-02-21 c7a6f88
Update MANIFEST.in
