+++ MANIFEST.in
@@ -1,3 +1,3 @@
 include *.txt
 include *.md
-include /VIIRS_Sample/*
+include CometTS/VIIRS_Sample/*
