+++ README.md
@@ -5,10 +5,10 @@
 <p align="center">
 <img align="center" src="https://img.shields.io/pypi/v/cometts.svg" alt="PyPI">
 <img align="center" src="https://img.shields.io/conda/vn/conda-forge/cometts.svg" alt="conda-forge">
-<img align="center" src="https://travis-ci.com/CosmiQ/cometts.svg?branch=master" alt="build">
-<img align="center" src="https://img.shields.io/github/license/cosmiq/cometts.svg" alt="license">
+<img align="center" src="https://travis-ci.com/jshermeyer/cometts.svg?branch=master" alt="build">
+<img align="center" src="https://img.shields.io/github/license/jshermeyer/cometts.svg" alt="license">
 <img align="center" src="https://img.shields.io/docker/build/cosmiqworks/cometts.svg" alt="docker">
-<a href="https://codecov.io/gh/CosmiQ/cometts"><img align="center" src="https://codecov.io/gh/CosmiQ/cometts/branch/master/graph/badge.svg" /></a>
+<a href="https://codecov.io/gh/jshermeyer/cometts"><img align="center" src="https://codecov.io/gh/jshermeyer/cometts/branch/master/graph/badge.svg" /></a>
 </p>
 
 - [Installation Instructions](#installation-instructions)
