### 2019-02-21 1139a1e
Delete Comet_test.py
