+++ .travis.yml
@@ -30,6 +30,7 @@ install:
   - pip install -q -e .
   - pip install codecov pytest pytest-cov
 # command to run tests
+before_script: cd CometTS
 script:
   - pytest --cov=./
 
