+++ .travis.yml
@@ -32,7 +32,7 @@ install:
 # command to run tests
 before_script: cd CometTS
 script:
-  - pytest --cov=./
+  - python3 -m pytest --cov=./
 
 after_success:
   - codecov
