+++ .travis.yml
@@ -32,7 +32,7 @@ install:
 # command to run tests
 before_script: cd CometTS
 script:
-  - python3 -m pytest --cov=./
+  - python -m pytest --cov=./
 
 after_success:
   - codecov
