+++ MANIFEST.in
@@ -1,3 +1,5 @@
 include *.txt
 include *.md
 include CometTS/VIIRS_Sample/*
+include CometTS/tests/*
+include CometTS/Landsat_Expansion_Docs_HelperScripts/*
