### 2019-02-23 c0b91b9
Added functionality to specify output locations.
