### 2019-02-23 a50a7c7
Added functionality to specify output locations.
