### 2019-02-23 1259433
Delete ARIMA.py
