### 2019-02-23 d2d659d
Delete CometTS.py
