+++ README.md
@@ -6,7 +6,7 @@
 <img align="center" src="https://img.shields.io/pypi/v/cometts.svg" alt="PyPI">
 <img align="center" src="https://travis-ci.com/jshermeyer/CometTS.svg?branch=master" alt="build">
 <img align="center" src="https://img.shields.io/github/license/jshermeyer/cometts.svg" alt="license">
-<img align="center" src="https://img.shields.io/docker/build/jshermeyer/cometts.svg" alt="docker">
+<img align="center" src="https://img.shields.io/docker/build/jss5102/cometts.svg" alt="docker">
 <a href="https://codecov.io/gh/jshermeyer/cometts"><img align="center" src="https://codecov.io/gh/jshermeyer/cometts/branch/master/graph/badge.svg" /></a>
 </p>
 
