### 2019-02-25 8c54b3a
Delete Dockerfile
