+++ Dockerfile/Dockerfile
@@ -1,27 +1,0 @@
-FROM python:3.6-stretch
-
-MAINTAINER jshermeyer
-
-# Update base container install
-RUN apt-get update
-RUN apt-get upgrade -y
-
-# Add unstable repo to allow us to access latest GDAL builds
-RUN echo deb http://ftp.uk.debian.org/debian unstable main contrib non-free >> /etc/apt/sources.list
-RUN apt-get update
-
-# Existing binutils causes a dependency conflict, correct version will be installed when GDAL gets intalled
-RUN apt-get remove -y binutils
-
-# Install GDAL dependencies
-RUN apt-get -t unstable install -y libgdal-dev g++
-
-# Update C env vars so compiler can find gdal
-ENV CPLUS_INCLUDE_PATH=/usr/include/gdal
-ENV C_INCLUDE_PATH=/usr/include/gdal
-
-# This will install latest version of GDAL
-RUN pip install GDAL>=2.2.4
-
-#Install Comet
-RUN pip3 install CometTS
