+++ README.md
@@ -8,7 +8,7 @@
 <img align="center" src="https://anaconda.org/jshermeyer/cometts/badges/version.svg" alt="conda">   
 <img align="center" src="https://travis-ci.com/jshermeyer/CometTS.svg?branch=master" alt="build">
 <img align="center" src="https://img.shields.io/github/license/jshermeyer/cometts.svg" alt="license">
-<img align="center" src="https://img.shields.io/docker/build/jss5102/:cometts.svg" alt="docker">
+<img align="center" src="https://img.shields.io/docker/build/jss5102/cometts.svg" alt="docker">
 <a href="https://codecov.io/gh/jshermeyer/cometts"><img align="center" src="https://codecov.io/gh/jshermeyer/cometts/branch/master/graph/badge.svg" /></a>
 </p>
 
