+++ README.md
@@ -18,25 +18,56 @@
 ---
 
 
+## Base Functionality
+The Comet Time Series tool is an open source jupyter notebook and command line utility that enables users to visualize or extract relevant statistics from a time series of overhead imagery within a specific area. A user can input a time series raster dataset and a polygon vector file (i.e., shapefile or geojson) that contains an area or set of areas. Within each area, a .csv file will be output containing relevant statistics. A plot will also be produced to enable simple yet effective visualizations.  Following a base analysis, CometTS enables anomaly detection using a user-specified historical baseline of time series imagery, and monthly Auto-Regressive Integrated Moving Average (ARIMA) Analysis.  Detailed workflows are described below.
 
-## About
-The Comet Time Series tool is an open source IPython Notebook deployment that enables users to visualize a time series of satellite or potentially airborne imagery within a specific area. A user can input a time series raster dataset and a polygon vector file (i.e., shapefile or geojson) that contains an area or set of areas. Within each area, a .csv file will be output containing relevant statistics. A plot will also be produced to enable simple yet effective visualizations.
-
-The first dataset we tested this tool with was [Suomii NPP VIIRS Nightlights Monthly Composite Data](https://ngdc.noaa.gov/eog/viirs/download_dnb_composites.html)  These data can be scraped easily using Firefox and the [DownThemAll](https://addons.mozilla.org/en-US/firefox/addon/downthemall/) AddOn. After download, the imagery will simply need to be unzipped using your tool of choice (i.e. WinRar or Mac Archive Utility) and formatted as described below.
-
-Note that these data are incredibly large in size and may take considerable time (e.g., over a day) to download the whole dataset for the entire earth (i.e., ~1TB of data). Another option is downloading just 1/6th of the data for the earth.  Uncompressed, this will be about 200GB.
 
 Check out the expansion pack & users guide for working with Landsat imagery [here.](https://github.com/CosmiQ/CometTS/blob/master/Landsat_Expansion_Docs_HelperScripts/Landsat_Expansion_CometTS_UsersGuide.docx)
 
 
+
 ## Installation
-Python 2.7+, GDAL, and Jupyter Notebooks are the baseline requirements. The other requirements are listed in Requirements.txt and are easily installable via pip or conda.
+Python 2.7 or 3.6 are the base requirements plus several packages.  CometTS can be installed in multiple ways including conda, pip, docker, and cloning this repository.
 
 
-    pip install -r Requirements.txt
+### Clone it
+We recommend cloning to add all sample data and easier access to the jupyter notebooks that leverage our plotting functions.
+```
+git clone https://github.com/CosmiQ/CometTS.git
+```
+If you would like the full functionality of a python package we have several options.
 
+### pip
+```
+pip install CometTS
+```
+pip installs may fail on macs with python3 as GDAL is finicky.  Use some of the alternative approaches below.
 
-## Dataset Organization
+### Docker
+```
+docker pull jss5102/cometts
+cd CometTS
+docker build .
+docker start CometTS
+```
+
+### Conda
+Create a conda environment!
+
+```
+git clone https://github.com/CosmiQ/CometTS.git
+cd CometTS
+conda env create -f environment.yml
+source activate CometTS
+```
+
+### Dependencies
+All dependencies can be found in the docker file [Dockerfile](./Dockerfile) or
+[environment.yml](./environment.yml) or [requirements.txt](./requirements.txt).
+
+## Dataset Acquisition & Organization
[…]
