+++ Dockerfile
@@ -21,7 +21,7 @@ ENV CPLUS_INCLUDE_PATH=/usr/include/gdal
 ENV C_INCLUDE_PATH=/usr/include/gdal
 
 # This will install latest version of GDAL
-RUN pip install GDAL>=2.2.4
+RUN pip install GDAL>=2.4
 
 #Install Comet
 RUN pip3 install CometTS
