### 2019-02-25 4e429bb
Update MANIFEST.in
