+++ environment.yml
@@ -1,7 +1,7 @@
 name: CometTS
 channels:
-  - defaults
   - conda-forge
+  - defaults
 dependencies:
   - numpy=1.15.4
   - pandas=0.24.1
@@ -11,7 +11,7 @@ dependencies:
   - fiona=1.8.4
   - affine=2.2.2
   - rasterstats=0.13.0
-  - matplotlib
+  - matplotlib=3.0.2
   - seaborn=0.9.0
   - jupyter
   - ipywidgets=7.4.2
@@ -19,5 +19,3 @@ dependencies:
   - tqdm=4.31.1
   - scipy=1.2.0
   - gdal=2.4.0
-  - pip:
-    - cometTS
