+++ Dockerfile
@@ -24,4 +24,4 @@ ENV C_INCLUDE_PATH=/usr/include/gdal
 RUN pip install GDAL>=2.4
 
 #Install Comet
-RUN pip3 install CometTS
+RUN pip3 install CometTS --upgrade
