+++ README.md
@@ -46,9 +46,7 @@ pip installs may fail on macs with python3 as GDAL is finicky.  Use some of the 
 ### Docker
 ```
 docker pull jss5102/cometts
-cd CometTS
-docker build .
-docker start CometTS
+docker run -it -v /nfs:/nfs --name cometts jss5102/cometts /bin/bash 
 ```
 
 ### Conda
