### 2019-03-20 #1047 @Kevin-Mattheus-Moerman
@jshermeyer in particular can you comment on whether you agree with @zhampel to migrate "... the functionality of your python notebooks to a more API-like structure..."?
