### 2019-05-13 #1047 @jshermeyer
> @jshermeyer here are my comments regarding the structure of CometTS's code. I found your instructions easy to follow with the tools you have built. Your code does work as advertised, though I do think there are a changes that should be made to the formatting of the code, including commenting and other coding standards that should be upheld. Furthermore, I think your work could be made much stronger as an extensible project, by migrating the functionality of your python notebooks to a more API-like structure. I would also like to point out that a significant hurdle to properly reviewing the project is the lack of easy access to example datasets (you do point this out), which is beyond the scope of you as the author. To this end, the open source community encourages steps like CometTS to making these tools openly available indeed.

Thanks to Zig @zhampel  for his helpful and thoughtful comments.  Also thanks to the editors @Kevin-Mattheus-Moerman @labarba  for pushing this on, I apologize for being an absentee on this.  I've gradually added onto Comet to address all of his concerns.

_A brief overview:_
-Code totally reformatted to meet pep8 standards and reorganized for simplicity
-Now has a command line API for those who want to ignore the notebooks (would still recommend notebooks for easy plotting/viz).
-Installable via pip/conda
-Dockerized
-Pytests included for sensitive functions
-Added support for autoregressive integrated moving average modeling for future trend forecasting
-Includes an example test dataset of NPP VIIRS monthly composites over San Juan, PR.

Specifics:

> MacOS fails to pip install due to gdal. Needs brew install gdal prior to pip install -r Requirements.txt. Works on my MacOS Mojave laptop.

Updated to have gdal install at the end.  Also include support with docker and a conda environment for mac users.  Would not recommend brew installing gdal, things get ugly fast when using brew/pip/conda.

> DownThemAll add-on link points to a non-existent page.

Still works for old firefox, unfortunately isn't supported any more.  I'll leave it up to an end user to get their own data.  I now include sample VIIRS imagery for play.

> Ensuring a proper directory structure is clear for the example you gave, but is that the case with other datasets? Not being a geospatial data expert, is there a resource that provides guidance on proper structuring, or perhaps a universal industry standard?

This is a standard for a few time-series methods, i.e. CCDC/YATSM- https://github.com/ceholden/yatsm

> Had difficulty with GDAL install (who doesn’t?). Perhaps best to include gdal and subsequent deps on gdal at the end of the requirements list, so that all other packages install first. This may assist in debugging install.

Done.

> Also recommend including link to building GDAL with python-bindings in Installation section. Had success using following: http://trac.osgeo.org/gdal/wiki/BuildingOnUnix

Multiple choices for installing GDAL now available.

> Please include a python style .gitignore file that excludes the CometTS python notebooks.

Notebooks are pretty core for this project, but project is now pip/conda installable with an option for CLI only so this seems redundant.

> Code Specifics

> Recommend reformatting for python3 compatibility

Done.

> Recommended to import functions explicitly, instead of * for ease of tracking

Done.

> Header description for each fn/ Inputs and outputs description

Now have comments that describe inputs/outputs + github readme

> All input argument declarations should be initialized to None, ‘’, etc

Some of these I like to keep to give a user what an example input would look like.

> Run flake8 to format code properly, will make it easier to read too ;)

Done

> Multi_studyAreas_FullStats.csv is in same dir as notebooks. Plot_Results.ipynb looks elsewhere for it. Consider moving it into the example data dir. Yet, when trying to use this example csv data, I get an error that there is no key count. Did work on my generated csv data though.

Code structure has been totally reorganized

> For plotting, should have more scalable y-axis, as I chose a rather small city (Santa Fe, NM) to test

I've included a few helper notebooks for plotting only.  A user will have to have a bit of background with matplotlib for tweaking some of the plotting features.

> Difficult to debug errors, esp wrt formatting data issues

Not really sure how to make this easier, probably will require more iterations.  Have added pytests to hopefully alleviate some stress.

> What are the temp*.vrt files? Are they part of the project?

Temp outputs from some functions, can be ignored.  I likely should clean these up but could be helpful for debug.

> Need a set of tests for function verifications

Included for a good chunk of the code that could be sensitive.  Also include test data.
