### 2019-05-20 #1047 @dmittman
I find the material just outside my area of expertise, being more about satellite imagery analysis with a time-series component than pure time-series visualization.
