### 2019-05-21 #1047 @danielskatz
@turmon, @darth-pr - would either of you be able to step in and help with a review of this submission & this software for JOSS?  (Or perhaps suggest someone else who might?)

See [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) in case you are not aware of how JOSS works...
