### 2019-05-21 #1047 @turmon
Maybe Kyo Lee in my group?  He's a climate/remote sensing model analysis person, who contributes to an open source regional climate model system.

https://orcid.org/0000-0003-3754-3204

Huikyo.Lee@jpl.nasa.gov

Mike

> On May 20, 2019, at 6:24 PM, Daniel S. Katz <notifications@github.com> wrote:
>
> @turmon, @darth-pr - would either of you be able to step in and help with a review of this submission & this software for JOSS? (Or perhaps suggest someone else who might?)
>
> See JOSS review criteria in case you are not aware of how JOSS works...
>
