### 2019-05-21 #1047 @turmon
I was not able to find a GH account, but I can't rule it out.
Mike

> On May 21, 2019, at 8:44 AM, Daniel S. Katz <notifications@github.com> wrote:
>
> Thanks Mike - Does Kyo has a github account? I'll let @Kevin-Mattheus-Moerman follow-up with him, as the editor.
>
