### 2019-05-22 #1047 @Kevin-Mattheus-Moerman
@lubiavinhas @mloning @TonyBagnall @gilbertocamara @rolfsimoes @pedro-andrade-inpe @ammaciel @jgrss @slandersson @mmann1123 @avanetten @aftimg  would you be interested in reviewing this submission (about time-series analysis of satellite image data) for the [Journal of Open Source Software (JOSS)](http://joss.theoj.org/)? The [review process](https://joss.readthedocs.io/en/latest/review_criteria.html), which takes place in this GitHub issue, focuses largely on [the software](https://github.com/CosmiQ/CometTS) and a [short paper](https://github.com/openjournals/joss-papers/blob/joss.00866/joss.00866/10.21105.joss.00866.pdf).
Let me know if you have questions.
