### 2019-06-10 #1047 @Kevin-Mattheus-Moerman
@rmsare thanks for your help. I've created a set of tick boxes at the top of this issue which will guide you through the review process. Please also check out our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Let me know if you have any questions.
