### 2019-06-12 CosmiQ/CometTS#6 @rmsare
Make GitHub release and link version to Zenodo and paper
(This issue is part of a review of [the CometTS submission to JOSS](https://github.com/openjournals/joss-reviews/issues/1047). I will add a comment in that thread with a list of all items from my review when I've finished.)

The Zenodo archive should correspond to a [tagged release](https://github.com/CosmiQ/CometTS/releases). Currently the latest un-tagged source code is archived on Zenodo.

At the end of the review process, I'd suggest making a release on GitHub and updating the Zenodo archive and paper to reflect this.
