### 2019-06-12 CosmiQ/CometTS#7 @rmsare
Remove or clarify commented-out code
It's nice to see the tests pass, but there are a few instances of commented-out code in your tests and the example notebooks. You should remove these if they aren't necessary, or clarify if they can be run in certain cases (like you do in [ARIMA.py](https://github.com/CosmiQ/CometTS/blob/master/CometTS/ARIMA.py#L80) and elsewhere).

- [x] Remove old code and print statements in `ARIMA.py`: [L45-46](https://github.com/CosmiQ/CometTS/blob/master/CometTS/ARIMA.py#L46), [L120](https://github.com/CosmiQ/CometTS/blob/master/CometTS/ARIMA.py#L120), [L155-160](https://github.com/CosmiQ/CometTS/blob/4645553debe5cb8a59d0cd462d533a282221df5d/CometTS/ARIMA.py#L155), [L204-205, 210, 235, 237](https://github.com/CosmiQ/CometTS/blob/master/CometTS/ARIMA.py#L204)
- [x] Remove old code from `CometTS.py`. There are many in the plot functions like [L139](https://github.com/CosmiQ/CometTS/blob/master/CometTS/CometTS.py#L139), [L291-296](https://github.com/CosmiQ/CometTS/blob/master/CometTS/CometTS.py#L291), [L357 and 375](https://github.com/CosmiQ/CometTS/blob/master/CometTS/CometTS.py#L357), [L476-481](https://github.com/CosmiQ/CometTS/blob/master/CometTS/CometTS.py#L476), as well as old testing calls like the NDVI plot at [L847](https://github.com/CosmiQ/CometTS/blob/master/CometTS/CometTS.py#L847).
- [x] Remove old code from tests. It looks like you rewrote the code to use `pd.testing` instead of standard assertions, but many commented-out assertions remain (e.g. [line 62-65](https://github.com/CosmiQ/CometTS/blob/master/CometTS/tests/Comet_test.py#L62) and 67 of `Comet_test.py`). There's also a block-commented code snippet at the end of file that should be deleted.

This will make the examples more accessible and make it easier to extend or contribute to CometTS without wading through scratch code.

(You also might consider removing the statement "No plotting testing included, do your own damn plotting if you have a problem." I get that it's unrealistic to test all possible uses of bokeh and matplotlib in the CometTS workflow, but this could be a little off-putting to someone who might want to contribute.)

([JOSS review thread](https://github.com/openjournals/joss-reviews/issues/1047))
