### 2019-06-12 CosmiQ/CometTS#8 @rmsare
Correct paths to example data
When running the ARIMA_Plotting notebook, the user will have to edit the `input_CSV` variable to point to the relative path to your San Juan example. Otherwise the notebook will fail with an error.

I suggest that this should be specified as a relative path (`CometTS/VIIRS_Sample/San_Juan_ARIMA_Output.csv`), or alternatively include step-by-step instructions in each example notebook.

- [x] Change any absolute paths to relative paths in example notebooks

([JOSS review thread](https://github.com/openjournals/joss-reviews/issues/1047))
