### 2019-06-13 CosmiQ/CometTS#9 @rmsare
Add instructions for running tests
I suggest that brief instructions to install pytest and run it on `CometTS/tests` be added to either the README or CONTRIBUTING documents. It's not in `requirements.txt` so won't necessarily be in the development environment.

This will help contributors verify their changes don't break your codebase and help users determine what works if they install from a development branch.

- [x] Add testing instructions

([JOSS review thread](https://github.com/openjournals/joss-reviews/issues/1047))
