### 2019-06-13 CosmiQ/CometTS#10 @rmsare
Add doc strings for major functions; reformat function names per PEP 8
Every significant function should include [a simple docstring](http://google.github.io/styleguide/pyguide.html#383-functions-and-methods) to describe what it does, its arguments, and any outputs.

Related to this, I suggest renaming your functions according to [conventional Python style](https://www.python.org/dev/peps/pep-0008/#function-and-variable-names). For example, `Do_Zonal_Stats` would become `do_zonal_stats`. Using more descriptive names, like `calculate_zonal_stats`, would be even better. This should be easy to do across the package and notebooks by find and replace or `sed`.

As noted in @zhampel's review, it would be good to run a format checker like `flake8` on your codebase which would catch issues like this. It looks like you did this in your response, but there are still [some small whitespace issues](https://github.com/CosmiQ/CometTS/blob/master/CometTS/CometTS.py#L187).

- [x] Add docstrings/function headers
- [x] Reformat function names consistently

([JOSS review thread](https://github.com/openjournals/joss-reviews/issues/1047))
