### 2019-06-13 #1047 @rmsare
@jshermeyer I am reviewing your submission, and I had a question about your changes to the project.

I would like to suggest some reorganization, but it looks like a bit of that has been implemented in @zhampel's [fork of CometTS](https://github.com/zhampel/CometTS). Do you plan to incorporate any of these changes, e.g. to directory structure?

Thanks for clarifying - it will help us to provide a more manageable review.
