### 2019-06-14 CosmiQ/CometTS#6 @Kevin-Mattheus-Moerman
Yes creating an up to date archive of the reviewed software will be part of the general process.
