### 2019-06-14 CosmiQ/CometTS#11 @rmsare
Improve notebooks
The notebooks run smoothly, thank you for providing them. Here are some small changes that could make them much more user-friendly.

Each notebook should have a title and one sentence explaining what it does. Ideally some inline Markdown cells should tell the user what to change and what's going on.

Instead of a single script pasted in the first cell, I suggest you separate the cells logically so the user can step through the analysis. This might be: one cell for imports, one cell for user-defined variables not specified in a widget, one cell per widget definition, and a final "run" cell. Or, if you don't want to use cells, please add header comments as you did in `CSV_Creator.ipynb`.

See examples of clean notebooks and example scripts [here](https://examples.dask.org/applications/satellite-imagery-geotiff.html) and [here](https://www.fatiando.org/verde/latest/gallery/vector_uncoupled.html#sphx-glr-gallery-vector-uncoupled-py).

Finally, please populate the widget fields in `CometTS_Visualizer.ipynb` with relative paths to your San Juan example. This will allow the user to immediately see what CometTS can do.

- [x] Add headers and explanatory text to notebooks
- [x] Separate cells or add header comments for clarity
- [x] Populate default example in `CometTS_Visualizer.ipynb`

([JOSS review thread](https://github.com/openjournals/joss-reviews/issues/1047))
