### 2019-06-14 CosmiQ/CometTS#12 @rmsare
Condense example in README and link to separate documentation
Your README is a first entry point to the project, and should summarize what it does, how to contribute and get help, and where to find more details. Currently I feel that the detailed workflow makes it difficult to get a overall picture of the package.

- [x] Migrate README walkthrough to notebooks, or condense example
- [x] Provide easily-digestible workflow steps

Any examples in the README should be pretty short; it's better to give more detail in separate documentation that you link to. See [the verde package's README](https://github.com/fatiando/verde) and [documentation](https://www.fatiando.org/verde/latest/) for a good example (published in JOSS!).

I suggest that you move most of the Workflow example section to a separate notebook or notebooks, and provide only a short walkthrough of the key steps. Something like this, with the links pointing to the more detailed documentation:

## CometTS workflow

To use CometTS, you must define your area of interest, provide a CSV file documenting how your imagery is organized, and then run one of the CometTS analysis tools. This usually takes the following steps

1. Outline and download your AOI with a service like [geojson.io]()
2. Organize your imagery and document it with the [CometTS.CSV_It tool]()
3. Analyze your data using:
   - [CometTS]() for trend analysis  (optionally, mask unwanted clouds and other features with [`--maskit` option]())
   - or [CometTS.ARIMA]() for averaging and anomaly detection
4. Plot the results using one of [the plotting notebooks]()

([JOSS review thread](https://github.com/openjournals/joss-reviews/issues/1047))
