### 2019-06-14 #1047 @rmsare
Thanks for submitting this, @jshermeyer! Here is my initial review.

Overall, it looks like a useful set of tools for quick time series analysis of satellite imagery that may be appropriate for JOSS. Nice work.

While the code functions as advertised, the codebase and documentation could be improved with substantial changes that will make it easier to use. I believe implementing many of these changes will ensure that CometTS is a usable, extensible software package rather than an analysis utility.

I was able to install in a conda environment with PyPI.

The code is well-commented, but function-level documentation is not quite sufficient (see [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#api-documentation)). I also think the README could be streamlined and the example notebooks improved with some instructions. Someone interested in CometTS should be able to scan the example notebooks and documentation to quickly see how things work.

- [ ] Improve the README and add separate documentation ([#12](https://github.com/CosmiQ/CometTS/issues/12))
- [ ] Make notebooks more user friendly ([#11](https://github.com/CosmiQ/CometTS/issues/11))
- [ ] Add doc strings ([#10](https://github.com/CosmiQ/CometTS/issues/10))
- [ ] Document the testing protocol ([#9](https://github.com/CosmiQ/CometTS/issues/9))

Once you add docstrings you might consider [auto-generating API documentation with Sphinx and ReadtheDocs](https://sphinx-rtd-tutorial.readthedocs.io/en/latest/sphinx-quickstart.html) (file an issue with the `docs` tag?).

- The main notebooks are functional. They don’t run under Jupyter lab due to changes to the widgets extension, but they do run as standalone notebooks.
- The command line utilities are functional and run with real data.
- I agree with @zhampel that refactoring to an API design would be a good idea. This might be a good thing to plan out in an issue with the `refactor` or `enhancement` tags.

The manuscript is written well, but describes the CometTS workflow rather than making a case for the research software contribution. Consider adding an introduction with two or three sentences providing more context on satellite timeseries analysis and why CometTS is a good alternative to other open-source offerings.

Many of the details in the final paragraph could be summarized in a shorter sentence like *“CometTS output includes user-specified statistics such as mean, median, and quantiles, and the package offers masking functionality to remove clouds and snow from the area of interest.”* Then the remaining space could describe research use cases in more detail.

The [statement of need](https://joss.readthedocs.io/en/latest/review_criteria.html#a-statement-of-need) should state the intended research audience, which appears to include social scientists, earth scientists, and/or geospatial intelligence analysts.

- [ ] Add audience to statement of need
- [ ] Expand introduction, including appropriate citations to related work
- [ ] Edit manuscript to remove workflow details in favor of suggested research use cases
- [ ] Remove ® symbol from “Python ®”

### Minor issues

- [ ] **Cleanup:** See issues [#7](https://github.com/CosmiQ/CometTS/issues/7), [#8](https://github.com/CosmiQ/CometTS/issues/8)
- [ ] **Version:** The Zenodo archive and paper should correspond to a [tagged release](https://github.com/CosmiQ/CometTS/releases) on Github ([#6](https://github.com/CosmiQ/CometTS/issues/6))

**Recommendation:** Could be accepted after moderate enhancements
