### 2019-06-15 #1047 @jjmcnelis
> @jjmcnelis thanks for joining as reviewer. Would you be able to update your GitHub profile e.g. such that it lists your affiliation (and/or a link ORCID profile if possible). Thanks. Let me know if you have questions.

@Kevin-Mattheus-Moerman Happy to help. ORCID and Org are now listed on my profile. I'll complete my initial review before the business week opens on Monday. Please let me know if that doesn't agree with your expected schedule for reviewers and I'll up the urgency.
