### 2019-06-16 #1047 @jshermeyer
Thanks @rmsare for the thorough review.  Good recommendations that I'm going to gradually start addressing.  I don't have as much time for this repo as I used to as this project was more of a focus about a year ago.  Will see how it goes, may take some time.

-Jake
