### 2019-06-16 0d8626b
Added docstrings and reformated
