+++ CometTS/CometTS.py
@@ -1,21 +1,50 @@
 import os
-import glob
 import gdal
 import rasterio
 from rasterstats import zonal_stats
 import numpy as np
 import pandas as pd
 import geopandas as gpd
-import shapely.wkt
-import matplotlib.pyplot as plt
 from tqdm import tqdm as tqdm
-import matplotlib.dates as mdates
-from scipy.signal import gaussian
-from scipy.ndimage import filters
 import argparse
 
 
-def Process_imagery(Path_dir, zonalpoly, NoDataValue, mask_value, maskit=True, Path_out=""):
+def run_comet(directory_csv, zonalpoly, NoDataValue, mask_value, maskit=True, Path_out=""):
+    """Run CometTS.  Analyze your timeseries of raster data for your polygon(s) of interest.
+
+    Arguments
+    ---------
+    directory_csv : str
+        The specific path to the csv documenting your raster data and
+        directory structure created with CometTS.csv_it.
+    zonalpoly : str
+        The specific path to the polygon (geojson or shapefile or other vector)
+        that contains your areas of interest for analysis.
+    NoDataValue : int
+        The value of blank space where no actual data resides in an image. For
+        example Landsat surface reflenctance imagery has a value of -9999 where
+        no observations exist. Suggest a value of -1 for NPP VIIRS.  If your
+        imagery does not have a NoDataValue set == -1.
+    mask_value : int(s)
+        The value of a cloud or other anomaly mask (ex: Clouds/Cloud Shadow=1).
+         Multiple values can be used and be split by commas (i.e. 1,2,99) if no
+          masking is required, leave blank.
+    maskit : bool
+        Should cloud or anomalies be masked? Defaults to yes (``True``). If true
+        the function will use the mask_value(s) passed to automatically remove
+        any anomalies from the time series of imagery.
+    Path_out :str
+        The output path for detailed statistics of each polygon analyzed for the
+        entire time series of imagery.  All stats will be stored in csv format.
+        Defaults to the same path as directory_csv.
+
+    Returns
+    -------
+    gdf2 : a :class:`geopandas.geodataframe` that conains all statstics for all
+    ploygons and all imagery in the time series. Additonally all statistics for
+    each individual polygon will be output in csv format to the Path_out directory.
+    """
+
     print(zonalpoly)
     gdf = gpd.read_file(zonalpoly)
     if maskit:
@@ -23,9 +52,9 @@ def Process_imagery(Path_dir, zonalpoly, NoDataValue, mask_value, maskit=True, P
         mask_value = mask_value.split(",")
     # Get the zonal stats
     NoDataValue = int(NoDataValue)
-    gdf2 = Do_Zonal_Stats(Path_dir, gdf, NoDataValue, mask_value, maskit)
+    gdf2 = calculate_zonal_stats(directory_csv, gdf, NoDataValue, mask_value, maskit)
     # Get number of observations
-    gdf3 = Get_Num_Obs(Path_dir, gdf, NoDataValue, mask_value, maskit)
+    gdf3 = get_num_obs(directory_csv, gdf, NoDataValue, mask_value, maskit)
     for date in gdf2.iterrows():
         if 'median' in gdf3.columns:
             gdf2['observations'] = gdf3['median']
@@ -36,7 +65,7 @@ def Process_imagery(Path_dir, zonalpoly, NoDataValue, mask_value, maskit=True, P
     z_simple = z_simple[-1].split('.')
     z_simple = z_simple[0]
     if Path_out == "":
-        Path_out = os.path.dirname(os.path.abspath(Path_dir))
+        Path_out = os.path.dirname(os.path.abspath(directory_csv))
 
     output = os.path.join(Path_out, z_simple + '_FullStats.csv')
     print("CSV statistics saved here: ", output)
[…]
