+++ setup.py
@@ -34,6 +34,6 @@ setup(name='CometTS',
       install_requires=inst_reqs,
       extras_require=extra_reqs,
       entry_points={
-          'console_scripts': ['CometTS=CometTS.CometTS:main', 'CometTS.CSV_It=CometTS.CSV_It:main', 'CometTS.ARIMA=CometTS.ARIMA:main']
+          'console_scripts': ['CometTS=CometTS.CometTS:main', 'CometTS.csv_it=CometTS.csv_it:main', 'CometTS.ARIMA=CometTS.ARIMA:main']
       }
       )
