+++ CometTS/Plotting.py
@@ -1,0 +1,830 @@
+import numpy as np
+import pandas as pd
+import geopandas as gpd
+import shapely.wkt
+import matplotlib.pyplot as plt
+import matplotlib.dates as mdates
+from scipy.signal import gaussian
+from scipy.ndimage import filters
+
+"""This file contains multiple plotting functions.  Specific docstrings not
+provided but these are simple wrappers around matplotlib.  Simple adjustments to
+what you pass to each function should enable easy modification of output plots.
+"""
+
+
+def interpolate_gaps(values, limit=None):
+    """When there are gaps in the data they sometimes cannot be plotted properly.
+    This funciton interpolates these gaps using a linear interpolation to enable
+    plotting.
+
+    Arguments
+    ---------
+    values : a :class:`numpy.array`
+        The data that we want to remove the gaps within.
+
+    limit : int
+        The maximum amount of times to interpolate. i.e. if there is a large
+        gap in the data due to a lack of obserbations, how much should we
+         interpolate?  Defaults to None.
+
+    Returns
+    -------
+    filled : a :class:`numpy.array`
+        An array where any gaps in the data are removed via linear interpolation.
+    """
+    values = np.asarray(values)
+    i = np.arange(values.size)
+    valid = np.isfinite(values)
+    filled = np.interp(i, i[valid], values[valid])
+
+    if limit is not None:
+        invalid = ~valid
+        for n in range(1, limit + 1):
+            invalid[:-n] &= invalid[n:]
+        filled[invalid] = np.nan
+
+    return filled
+
+
+def run_plot(
+        main_gdf,
+        figsize=(
+            12,
+            6),
+        y_val_alpha=1,
+        scatter_alpha=1,
+        error_alpha=0.2,
+        y_label="Brightness",
+        x_label="Date",
+        title_label="Brightness over time - ID: ",
+        figname='',
+        custom_x_axis=True,
+        show_grid=True,
+        show_legend=True,
+        min_count=0.5,
+        ymin=0,
+        ymax=50,
+        dynamic_max=False):
+    print("Plotting...")
+    for item in main_gdf['ID'].unique():
+        plt.style.use('fivethirtyeight')
+        fig, ax = plt.subplots(figsize=figsize)
+        gdf3 = main_gdf[(main_gdf.ID == item)]
+        title = title_label + str(item)
+        gdf3 = gdf3.sort_values(['date'])
+        x = gdf3['date']
+        xdate = x.astype('O')
+        xdate = mdates.date2num(xdate)
+        y = gdf3['median']
+        if dynamic_max:
+            ymax = gdf3['median'].max() + (gdf3['median'].max() * 0.5)
+        if 'observations' in gdf3.columns:
+            count = gdf3['observations']
+        err_plus = gdf3['percentile_75']
+        err_minus = gdf3['percentile_25']
+        # Set the min_count value as a value from 0 to 1 (0 to 100%)
+        ax.set_ylim([ymin, ymax])
+
+        # This will filter out observations where over n% of a polygon is
+        # masked out
+        if min_count > 0:
+            z = gdf3['count']
+            zmax = gdf3['count'].max()
+            z = z / zmax
+            xdate = xdate[z >= min_count]
+            y = y[z >= min_count]
+            if 'observations' in gdf3.columns:
+                count = count[z >= min_count]
+            err_plus = err_plus[z >= min_count]
+            err_minus = err_minus[z >= min_count]
+
+        # plot regression line, if desired
+        idx = np.isfinite(xdate) & np.isfinite(y)
+        p2 = np.poly1d(np.polyfit(xdate[idx], y[idx], 1))
+        ax.plot(
+            xdate,
+            p2(xdate),
+            '-',
[…]
