### 2019-06-16 064d8f5
Update Comet_test.py
