### 2019-06-16 16477c9
Removed commented out code
