### 2019-06-16 #1047 @jjmcnelis
@jshermeyer I enjoyed working through your example - never had a use for the VIIRS night lights product. Please give me a few days to edit this comment with the relevant issues.

**Key recommendations:**
1. Provide a more explicit acknowledgement of whether CometTS is
a. a flexible workflow for processing raster time series, or
b. a tool for users to leverage in their own workflows.
2. Focus on features that distinguish CometTS from other analytical tools for processing raster time series, e.g. [TIMESAT](http://web.nateko.lu.se/timesat/timesat.asp).
3. Are the use cases compatible with CometTS limited to "overhead imagery"? This comment is specific to the language that you use in the paper and documentation. More thoughts below.

*@Kevin-Mattheus-Moerman Is 1a is compatible with JOSS submission requirements?*

### General
* Version discrepancy: pypi version > 1.0
* Maybe add a .gitignore file: [__pycache__](https://github.com/CosmiQ/CometTS/tree/master/CometTS/__pycache__) [__pycache__](https://github.com/CosmiQ/CometTS/tree/master/CometTS/tests/__pycache__)
* I was able to install in a conda environment.
* I was not able to install in a Python 3 virtual environment because of the common GDAL path hang-ups.

### Functionality
* Add example implementations for more imagery sources. Greater variety will attract users. The VIIRS example is cool, but I believe it represents an edge case for most users of VIIRS imagery.
* Test and document the imagery sources and formats that are compatible with CometTS. This will help users to determine if CometTS meets their needs.

**What differentiates CometTS from other time series raster processing tools?**

There are many tools on GitHub for processing raster time series. I've written a few elaborate workflows myself with the intention of extending them *someday*. CometTS needs more examples that demonstrate the variety of supported imagery to set itself apart.

One software that I've used in the past that provides similar functionality (not open-source, though) is [TIMESAT](http://web.nateko.lu.se/timesat/timesat.asp). The piece that sets TIMESAT apart is that it was designed for use with data with a seasonality component to it. Identify and promote things like that. The code that parses metadata from the GeoTIFF filenames is as important of a contribution as the zonal stats, in my view.

The UI crafted through the notebooks is clever. But they need more detail about inputs and outputs and which functions are called for them to be useful examples.

**Are raster formats other than GeoTIFF supported?**

The documentation suggests that the intended audience is users of satellite imagery, but many of the relevant datasets are distributed in other formats. VIIRS is distributed in HDF for most standard products, not GeoTIFF like the dataset in your example. The metadata stored internally by HDF and similar formats eliminates the need for a config file like the one generated by [CSV_lt.py](https://github.com/CosmiQ/CometTS/blob/master/CometTS/CSV_It.py). Support for other GDAL-supported formats will be straightforward to implement if you choose to refactor as an API as @zhampel and @rmsare recommend.

If GeoTIFF is the only supported format, and you don't plan to add support for other formats, acknowledge this limitation prominently in the README.md. Also, consider using language that indicates that CometTS will render the time series statistics for most raster datasets arranged within a uniform grid and spatial reference system, [not just overhead imagery as stated in the opening sentence of the README.md](https://github.com/CosmiQ/CometTS#base-functionality).

### Documentation
Nice that Docker, requirement.txt, and environment.yml files and the relevant install instructions are included.
* Acknowledge whether the CometTS is a robust, flexible workflow or tools for a user to build their own workflow.

My feeling is that you don't expect the user to interact directly with the functions that underly the workflow demonstrated in the README. If that assumption is correct, adding inline annotations to your code is important so users can tweak it to meet their needs. If you decide to create an API, adding/improving the docstrings is a must.

* As stated under Functionality, provide examples for more data sources. It's not clear how transferable the VIIRS workflow example is to other data sources (i.e. previous comment about VIIRS and HDF).
* Cut and paste the demonstration from the README into a notebook or notebooks, as recommended by @rmsare. The notebooks need more detail.

### Paper
My feelings about the content of the paper are pretty well covered above. I have no problems with the grammar/writing style but I can provide a marked up copy if you need that level of detail.
