### 2019-06-17 #1047 @Kevin-Mattheus-Moerman
@zhampel, @rmsare, @jjmcnelis I've paused this review for now since @jshermeyer has indicated that addressing the issue raised may take a significant amount of time.
@jjmcnelis thanks also for your review comments. You mentioned you might be able to open specific issues related to your comments. Please do, so @jshermeyer can work on these too.

Once @jshermeyer indicates they are ready to resume the review I'll remove the `paused` label and will call upon the reviewers again.
