### 2019-06-18 #1047 @jshermeyer
Thanks @jjmcnelis for another strong review.  I think this process will make a much stronger toolset when it's all said and done.

I've update the docstrings which I think was the most cumbersome thing to add. Next on my list is documenting the notebooks a bit better.  After that's adjusting the readme and responding to some of the more minor comments as proposed by reviewers.  Seems like it could be done in a day or two if I can steal a few hours.

My timing on this is quite ambiguous and  my availability to work on this will decline rapidly starting next week.  If I can get to it over the next 2 weeks I have a feeling I could get it done by then.  Otherwise this may bleed for awhile.

@Kevin-Mattheus-Moerman Am I correct in assuming no further reviewers will be added?  I'd hope to address everything in one shot and then move forward to a decision.
