+++ ORIGINAL_README.md
@@ -1,0 +1,260 @@
+<h1 align="center">Comet Time Series (CometTS) Visualizer</h1>
+
+![Niamey Time Series Plot](ExamplePlots/Niamey.png)
+
+<p align="center">
+
+<img align="center" src="https://img.shields.io/pypi/v/cometts.svg" alt="PyPI">
+<img align="center" src="https://anaconda.org/jshermeyer/cometts/badges/version.svg" alt="conda">   
+<img align="center" src="https://travis-ci.com/jshermeyer/CometTS.svg?branch=master" alt="build">
+<img align="center" src="https://img.shields.io/github/license/jshermeyer/cometts.svg" alt="license">
+<img align="center" src="https://img.shields.io/docker/build/cosmiqworks/cw-eval.svg" alt="docker">
+<a href="https://codecov.io/gh/jshermeyer/cometts"><img align="center" src="https://codecov.io/gh/jshermeyer/cometts/branch/master/graph/badge.svg" /></a>
+</p>
+
+- [Installation Instructions](#installation-instructions)
+- [Dependencies](#dependencies)
+- [License](#license)
+---
+
+
+## Base Functionality
+The Comet Time Series tool is an open source jupyter notebook and command line utility that enables users to visualize or extract relevant statistics from a time series of overhead imagery within a specific area. A user can input a time series raster dataset and a polygon vector file (i.e., shapefile or geojson) that contains an area or set of areas. Within each area, a .csv file will be output containing relevant statistics. A plot will also be produced to enable simple yet effective visualizations.  Following a base analysis, CometTS enables anomaly detection using a user-specified historical baseline of time series imagery, and Auto-Regressive Integrated Moving Average (ARIMA) Analysis.  Detailed workflows are described below.
+
+
+Check out the expansion pack & users guide for working with Landsat imagery [here.](https://github.com/CosmiQ/CometTS/blob/master/Landsat_Expansion_Docs_HelperScripts/Landsat_Expansion_CometTS_UsersGuide.docx)
+
+
+
+## Installation
+Python 2.7 or 3.6 are the base requirements plus several packages.  CometTS can be installed in multiple ways including conda, pip, docker, and cloning this repository.
+
+
+### Clone it
+We recommend cloning to add all sample data and easier access to the jupyter notebooks that leverage our plotting functions.
+```
+git clone https://github.com/CosmiQ/CometTS.git
+```
+If you would like the full functionality of a python package we have several options.
+
+### pip
+```
+pip install CometTS
+```
+pip installs may fail on macs with python3 as GDAL is finicky.  Use some of the alternative approaches below.
+
+### Docker
+```
+docker pull jss5102/cometts
+docker run -it -v /nfs:/nfs --name cometts jss5102/cometts /bin/bash 
+```
+
+### Conda
+Create a conda environment!
+
+```
+git clone https://github.com/CosmiQ/CometTS.git
+cd CometTS
+conda env create -f environment.yml
+source activate CometTS
+pip install CometTS
+```
+
+### Dependencies
+All dependencies can be found in the docker file [Dockerfile](./Dockerfile) or
+[environment.yml](./environment.yml) or [requirements.txt](./requirements.txt).
+
+## Dataset Acquisition & Organization
[…]
