+++ README.md
@@ -19,11 +19,14 @@
 
 
 ## Base Functionality
-The Comet Time Series tool is an open source jupyter notebook and command line utility that enables users to visualize or extract relevant statistics from a time series of overhead imagery within a specific area. A user can input a time series raster dataset and a polygon vector file (i.e., shapefile or geojson) that contains an area or set of areas. Within each area, a .csv file will be output containing relevant statistics. A plot will also be produced to enable simple yet effective visualizations.  Following a base analysis, CometTS enables anomaly detection using a user-specified historical baseline of time series imagery, and monthly Auto-Regressive Integrated Moving Average (ARIMA) Analysis.  Detailed workflows are described below.
+Comet Time Series (CometTS) is an open source tool coded in python including jupyter notebooks and command line utility that enables users to visualize or extract relevant statistics from a time series of overhead imagery within a specific area. To use CometTS, you must define your area of interest, provide a CSV file documenting how your imagery is organized, and then run one of the CometTS analysis tools. This usually takes the following steps
 
-
-Check out the expansion pack & users guide for working with Landsat imagery [here.](https://github.com/CosmiQ/CometTS/blob/master/Landsat_Expansion_Docs_HelperScripts/Landsat_Expansion_CometTS_UsersGuide.docx)
-
+1. Outline and download your AOI with a service like [geojson.io](www.geojson.io)
+2. Organize your imagery and document it with the [CometTS.CSV_It tool](https://github.com/CosmiQ/CometTS/blob/master/CometTS/CSV_It.py)
+3. Analyze your data using:
+   - [CometTS](https://github.com/CosmiQ/CometTS/blob/master/CometTS/CometTS.py) for trend analysis  (optionally, mask unwanted clouds and other features with [`--maskit` option](https://github.com/CosmiQ/CometTS/search?l=Python&q=--maskit))
+   - or [CometTS.ARIMA](https://github.com/CosmiQ/CometTS/blob/master/CometTS/ARIMA.py) for averaging and anomaly detection
+4. Plot the results using [the plotting notebooks](https://github.com/CosmiQ/CometTS/blob/master/Notebooks/Plot_Results.ipynb)
 
 
 ## Installation
@@ -64,166 +67,7 @@ pip install CometTS
 All dependencies can be found in the docker file [Dockerfile](./Dockerfile) or
 [environment.yml](./environment.yml) or [requirements.txt](./requirements.txt).
 
-## Dataset Acquisition & Organization
-The first dataset we tested this tool with was [Suomii NPP VIIRS Nightlights Monthly Composite Data](https://ngdc.noaa.gov/eog/viirs/download_dnb_composites.html). After download, the imagery will simply need to be unzipped using your tool of choice and formatted as described below. This repo also contains a limited amount of VIIRS sample data over San Juan, Puerto Rico.  Note if you are downloading from the NOAA website that these data are incredibly large in size and may take considerable time (e.g., over a day) to download the whole dataset for the entire earth (i.e., ~1TB of data). Another option is downloading just 1/6th of the data for the earth.  Uncompressed, this will be about 200GB.
-
-Data should be stored in a hierarchical folder pattern that is common for remote sensing applications.
-
-    VIIRSData/
-    └── Lat075N_Long060W
-        ├── SVDNB_npp_20120401-20120430_75N060W_vcmcfg_v10_c201605121456
-        │   ├── SVDNB_npp_20120401-20120430_75N060W_vcmcfg_v10_c201605121456.avg_rade9
[…]
