### 2019-06-21 a9a13c3
Update paper.md
