+++ paper.md
@@ -19,7 +19,7 @@ bibliography: paper.bib
 
 # Summary
 
-We have developed a tool that we called Comet Time Series (``CometTS``) that facilitates analysis and visualization of a time series of satellite imagery. The tool aims to enable population estimation research, land use change detection, or natural disaster monitoring using a range of data types. ``CometTS`` provides a partially automated approach for analyzing a time series of satellite images in any user defined polygon identifying a an arbitrarily sized region of interest (ROI). The tool calculates relevant statistical quantities (e.g., measures of central tendency and variation) and provides a visualization of their changes over time (Figure 1 and Figure 2). We believe this work is novel, as presently we are not aware of any such open-source tools to evaluate and leverage a time series of satellite or potentially airborne imagery from an ROI. Furthermore, this tool makes time series satellite imagery more accessible to the data science community and removes a Geographic Information Systems (GIS) tool as a requirement for working with these data. The tool requires only a web browser, Python, and dependent packages to function. Other time series tools like TSTools[@Holden:2017] are powerful but require a GIS interface and are limited to an analysis of individual pixels rather than larger areas of interest.
+We have developed a tool that we called Comet Time Series (``CometTS``) that enables workflows for the analysis and visualization of a time series of satellite imagery. The tool aims to enable population estimation research, land use change detection, or natural disaster monitoring using a range of data types. ``CometTS`` provides a partially automated approach for analyzing a time series of satellite images in any user defined polygon identifying a an arbitrarily sized region of interest (ROI). The tool calculates relevant statistical quantities (e.g., measures of central tendency and variation) and provides a visualization of their changes over time (Figure 1 and Figure 2). We believe this work is novel, as presently we are not aware of any such open-source tools to evaluate and leverage a time series of satellite or potentially airborne imagery from an ROI. Furthermore, this tool makes time series satellite imagery more accessible to the data science community and removes a Geographic Information Systems (GIS) tool as a requirement for working with these data. The tool requires only a web browser, Python, and dependent packages to function. Other time series tools like TSTools[@Holden:2017] are powerful but require a GIS interface and are limited to an analysis of individual pixels rather than larger areas of interest.
 
 ![Figure 1](https://raw.githubusercontent.com/CosmiQ/CometTS/master/ExamplePlots/Workflow.png)
 **Fig. 1:** A four step process to analyze and visualize trends in a time series of overhead imagery.
