+++ paper.bib
@@ -17,3 +17,35 @@
   doi          = {10.5281/zenodo.1324419},
   url          = {https://doi.org/10.5281/zenodo.1324419}
 }
+
+@article{Shermeyer_PR:2018,
+	doi = {10.1117/12.2325585},
+	url = {https://doi.org/10.1117/12.2325585},
+	year = 2018,
+	volume = {10790},
+	author = {Jacob Shermeyer},
+	title = {Assessment of electrical and infrastructure recovery in Puerto Rico following hurricane Maria using a multisource time series of satellite imagery},
+	journal = {SPIE Remote Sensing 2018 Conference Proceedings: Earth Resources and Environmental Remote Sensing/GIS Applications IX}
+}
+
+@article{Jonsson_Eklundh:2004,
+	doi = {10.1016/j.cageo.2004.05.006},
+	url = {https://doi.org/10.1016/j.cageo.2004.05.006},
+	year = 2004,
+	volume = {30},
+  issue = {8},
+	pages = {833--845},
+	author = {Per Jönsson and Lars Eklund},
+	title = {TIMESAT—a program for analyzing time-series of satellite sensor data},
+	journal = {Computers & Geosciences}
+}
+
+
+@misc{e-sensing/sits:2018,
+  author       = {Gilberto Camara; Rolf Simoes; Pedro R. Andrade; Victor Maus; alber sánchez; Luiz Fernando Ferreira Gomes de Assis; lorenalves; ywatacarvalho; Adeline Marinho Maciel; Lubia Vinhas; Gilberto Ribeiro de Queiroz},
+  title        = {e-sensing/SITS - Satellite Image Time Series Analysis: Version 1.12.5},
+  month        = dec,
+  year         = 2018,
+  doi          = {10.5281/zenodo.1974065},
+  url          = {https://doi.org/10.5281/zenodo.1974065}
+}
