+++ paper.bib
@@ -33,7 +33,7 @@
 	url = {https://doi.org/10.1016/j.cageo.2004.05.006},
 	year = 2004,
 	volume = {30},
-  issue = {8},
+  	issue = {8},
 	pages = {833--845},
 	author = {Per Jönsson and Lars Eklund},
 	title = {TIMESAT—a program for analyzing time-series of satellite sensor data},
