### 2019-06-24 cf2afc7
Update paper.md
