### 2019-06-24 ad63739
Update paper.md
