### 2019-06-24 #1047 @jshermeyer
@Kevin-Mattheus-Moerman, @zhampel, @rmsare, @jjmcnelis Thanks for the reviews and constructive feedback.  I've worked to incorporate all of your suggestions throughout this process into the repository in one way or another. As I stated previously I think the process has been worthwhile and I’m appreciative for your time to make CometTS a better and more usable package.

**What I've worked on**
Specific items that I've changed in the code, readme, or paper are all tracked in the issues: https://github.com/CosmiQ/CometTS/issues

Specifically issue 15 tracks the minor comments from all reviewers and aggregates them in one location.

**Specific responses**
A few things throughout the review likely require a direct response from me, I’ll run through them here.

> 1.	Acknowledge in the documentation whether CometTS is
> a. a workflow for processing/visualizing raster time series, or
> b. tools for workflows; i.e. it's unclear how many of the functions are reliant on outputs from other pieces.

I think CometTS is both a package with unique functionality,  a workflow, and provides tools for workflows for processing and visualizing raster time series.  So all of the above.

> 1.	Acknowledge the key features that distinguish CometTS from other processing/visualization tools for raster time series, e.g. TIMESAT.

Key differences are:
1)	the ability to handle arbitrarily sized areas of interest (polygons) vs. individual pixels for visualization.
2)	Anomaly detection and auto-regressive integrated moving average forecasting
3)	Python / jupyter notebook deployment

> Are raster formats other than GeoTIFF supported?
> I don't see any reference to other formats than GeoTIFF. Display a list of supported formats prominently on the README. Support for other GDAL-supported formats will be straightforward to implement if you choose to refactor as an API as zhampel and rmsare recommend. The metadata stored internally by HDF and others eliminates the need for a config file like the one generated by CSV_lt.py.
> •	Test and indicate in the README which imagery sources and formats are compatible with CometTS.
> The VIIRS example is cool, but I think it represents an edge case for most VIIRS users. VIIRS is distributed in HDF for most standard products, not GeoTIFF like in the example. The documentation suggests that users of "overhead imagery" are the intended audience, but many of the relevant datasets are distributed in other formats.

Have updated the readme to address this.  Any GDAL format raster is supported.  Any OGR format vector file is supported.  All metadata unfortunately is not created equally and no standard exists for what is contained in each file (although one is being worked on).  I think an ad-hoc approach to extracting at least the date information from each raster file is still the best method and lends flexibility to an end user. CometTS will auto-extract extent and projection if it is appropriately defined in the rasters.   HDF has no standard metadata format, and is a really a poor choice by NOAA to use.  In fact, some of their recent data has no projection or extent information embedded in each file. When you don’t have that from the start, you’re setting people up to fail.  If you have experience getting these data to play nicely out of the box with a GIS, I’d love to hear about it.

> •	Provide examples for more data sources. It's not clear how transferable the VIIRS workflow example is to other data sources.

This I really can’t do.  I’ve built the package, now it’s up to end users to run with it, and get creative with their scientific pursuits.  I’ve written a full paper regarding tracking electrification rates in Puerto Rico following Hurricane Maria using CometTS (now cited in the paper).  Additionally, I’ve shown separate examples with Landsat (transferable to any multi-spectral imagery with some tweaking) and using VIIRS for estimating seasonal population migrations.

**Future Things**
If/when the package is approved for JOSS, I intend to update the zenodo code as well as the code on pypi/conda.  Want to make sure there are no more modifications or recommendations before I do this.
