### 2019-06-26 #1047 @rmsare
@jshermeyer Thanks for your response. I will be able to look through your changes in more detail this weekend.

@Kevin-Mattheus-Moerman, I will add a comment here when that's done and follow up in review issues where appropriate.
