### 2019-07-04 CosmiQ/CometTS#7 @rmsare
Looks good - code has been cleaned up substantially
