### 2019-07-04 CosmiQ/CometTS#9 @rmsare
Instructions have been added in `CONTRIBUTING.md`
