### 2019-07-04 CosmiQ/CometTS#10 @rmsare
The new functions are docstrings are a good improvement. Thanks for updating.
