### 2019-07-04 CosmiQ/CometTS#11 @rmsare
The default example is now ready to run, and the notebooks have been improved
