### 2019-07-04 CosmiQ/CometTS#12 @rmsare
More detailed documentation has not been generated from the docstrings, but the clarity of the README has been improved.
