### 2019-07-04 CosmiQ/CometTS#16 @rmsare
Edit paper for style and clarity
The paper has been improved, but I feel like some detail is missing. The main issues I see are:

- [ ] The core contribution could be better defined in the second paragraph.

Right now, the paper states that CometTS provides new functionality in moving window statistics, extracting a time series from an ROI, and cloud masking. SITS, for one, has ARIMA functionality, and TIMESAT and SITS provide cloud/QA filtering, so I don't think these are the new, key features of CometTS. You should probably emphasize other unique advantages of CometTS as a rapid inspection workflow while mentioning these as important features.

- [ ] There is some redundant use of the first person. For instance,  you could edit the first sentence to remove "We have developed .. that we called" to more directly describe the software. This would be something like: "Comet Time Series (CometTS) is a set of Python notebooks that enables analysis and visualization of a time series of satellite imagery for the data science and geographic communities."

- [ ] Typo in the last paragraph of the paper ("calucluated" -> "calculated").
