### 2019-07-04 #1047 @rmsare
@Kevin-Mattheus-Moerman Looking through the changes, my major comments have been addressed by @jshermeyer and I feel like the notebooks and code have been made much more accessible.

Right now, I feel like the code and packaging meets the standards of JOSS, but it is up to the author to justify CometTS as a novel research tool. As @jjmcnelis asked above, could editors clarify when a workflow-oriented tool like this submission is defined as a "minor utility"?

The paper should be edited a bit ([#16](https://github.com/CosmiQ/CometTS/issues/16)). In particular, it claims CometTS provides "new" functionality (averaging, cloud masking, etc.) that is available in other open source software like SITS.
