### 2019-08-01 #1047 @Kevin-Mattheus-Moerman
@jjmcnelis can  you review the latest changes and @jshermeyer 's response to your points.

@jjmcnelis @rmsare To answer your question yes this package is considered more than a "minor utility", and it is clear it may be valuable for scientific research applications. A typical "minor utility" is a package doing something quite trivial or simple and which does not influence scientific research and the findings much or at all. Therefore one would likely not feel the need to cite such a minor utility. In this case however the package does influence the science as it provides both visualization and processing tools and workflows.
