### 2019-08-01 #1047 @Kevin-Mattheus-Moerman
@zhampel can you finalize the review on your end? Are you satisfied with the changes made? Thanks
