### 2019-08-09 #1047 @jjmcnelis
@Kevin-Mattheus-Moerman Apologies for the radio silence over the last week.

@jshermeyer I'm satisfied with the improvements to the repo. My main critiques about the README content have all been addressed:
1. Outlined basic steps for the primary use case (check - displayed prominently atop README.)
2. Supported formats (check - would've been fine with "All GDAL/OGR compatible formats are supported.", but it's nice to be reminded that I need to update my GDAL bookmarks :)
3. Replacing the details in the README with the three (?) additional time series examples better communicates your intent in developing this software. And the plots are fantastic. Clear and information rich, with concise descriptions of the phenomena captured by the data and depicted through CometTS.

Your updates to the paper help to clarify CometTS's place among the other tools that provide some similar functionality.

I apologize for not giving you more specific feedback during my initial review. Several commitments came calling at once. @rmsare Thanks for your diligence in documenting the review comments in the issues on CometTS repo!
