### 2019-08-09 #1047 @zhampel
@Kevin-Mattheus-Moerman I see that the author has addressed my earlier comments as well as those of the other reviewers. The documentation has been improved, and I think users will appreciate the addition of sample data for testing purposes. The final two boxes have been checked. Thanks.
