### 2019-08-09 #1047 @Kevin-Mattheus-Moerman
@jshermeyer @rmsare all boxes are ticked. However is it correct to say some work remains related to https://github.com/CosmiQ/CometTS/issues/16 raised by @rmsare?
