### 2019-08-09 #1047 @Kevin-Mattheus-Moerman
Thanks!

*Dr K.M. Moerman* <https://kevinmoerman.org/>
*Lecturer in Biomedical Engineering*

*Research Affiliate, MIT Media Lab*
+353 (0)91 49 3552
*ORCID iD* 0000-0003-3768-4269 <http://orcid.org/0000-0003-3768-4269>
Biomedical Engineering, Alice Perry Engineering Building
NUI Galway, University Road, Galway, H91 HX31, Ireland

On Fri, 9 Aug 2019 at 19:40, Zigfried Hampel-Arias <notifications@github.com>
wrote:

> @Kevin-Mattheus-Moerman <https://github.com/Kevin-Mattheus-Moerman> I see
> that the author has addressed my earlier comments as well as those of the
> other reviewers. The documentation has been improved, and I think users
> will appreciate the addition of sample data for testing purposes. The final
> two boxes have been checked. Thanks.
>
