### 2019-08-09 #1047 @rmsare
@Kevin-Mattheus-Moerman @jshermeyer Yes, I think the claims in the paper need to be revised a bit to describe what's new in CometTS, and some small edits remain. Otherwise the package is in good shape, and I'd be happy to recommend an updated submission for acceptance.
