+++ CometTS/CometTS.py
@@ -110,14 +110,14 @@ def mask_imagery(rasterin, mask, geom, NoDataValue, mask_value):
     MRO : a :class:`numpy.array` (masked raster object (MRO)). This is an array
     containing the original data that has been anomaly masked using the mask image.
 
-    affineO : a :class:`rastrio.affine????` an object that contains the specific
+    affineO : a :class:`rastrio.affine` an object that contains the specific
     location of the polygon of interest, indexed to the time series imagery.
     This is used to only read in a portion of the full raster instead of the
      full image in the future.  This improves computational performance.
     """
 
     geom = geom.bounds
-    #geom = [geom[0], geom[3], geom[2], geom[1]]
+    geom = [geom[0], geom[3], geom[2], geom[1]]
     MRO = gdal.Translate("temp1.vrt", rasterin, projWin=geom).ReadAsArray()
     MR2 = gdal.Translate("temp2.vrt", mask, projWin=geom).ReadAsArray()
     r = rasterio.open("temp1.vrt")
