### 2019-08-29 #1047 @jshermeyer
Hi Kevin,

@Kevin-Mattheus-Moerman <https://github.com/Kevin-Mattheus-Moerman> Updated
the text to make it seem as if "Cloud and snow masking" were not new
functionality.  If there is any further specific changes required for the
text let me know, otherwise I am satisfied with it.

-Jake

On Sat, Aug 24, 2019 at 5:30 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @jshermeyer <https://github.com/jshermeyer> can you give an update on
> progress in terms of replying/dealing with @rmsare
> <https://github.com/rmsare> comments? It looks like we are nearly there.
>
