### 2019-08-29 CosmiQ/CometTS#16 @rmsare
Thanks you for the edits. It's nearly there, but these issues remain in the text:

- [x] First paragraph: 'geographic' should be 'geography community' or 'geographic science community'
- [ ] Second paragraph: "adds new functionality to the ... domain" should be "includes functionality for overhead time series analysis and visualization", because this functionality is present in existing TSA packages.
- [x] Typo in the last paragraph of the paper ("calucluated" -> "calculated")
