### 2019-08-29 #1047 @rmsare
@jshermeyer, Thanks for making that change. It's nearly there, but some small edits are required for clarity and a typo.

I've noted the three most important changes to be made in a comment to [this issue](https://github.com/CosmiQ/CometTS/issues/16).
