### 2019-09-04 #1047 @Kevin-Mattheus-Moerman
Thanks @zhampel for your contribution. I've just unticked the "conflict of interest" box for you since you indicated to me in an email that "Jake Shermeyer, is a member of a sister lab of [your] organization". I'm reiterating that here for the record.
