### 2019-09-12 CosmiQ/CometTS#16 @Kevin-Mattheus-Moerman
@jshermeyer can you finalize these steps and report back over at https://github.com/openjournals/joss-reviews/issues/1047?
