### 2019-09-12 CosmiQ/CometTS#16 @jshermeyer
Apologies for the delay, I've been traveling/preoccupied.  I feel like today is the first time I've actually seen this issue for some reason so that could contribute to my tepid edits.  I've fixed two of the three issues noted by @rmsare. I'm pushing back on the second paragraph change, it presently reads:

> Specifically, ``CometTS`` output includes user-specified statistics such as mean, median, and quartiles, across arbitrarily sized regions of interest (ROI). Furthermore, the option for anomaly detection is also included, and ``CometTS`` leverages an Auto-Regressive Integrated Moving Average (ARIMA) analysis to quantify trends and test if observations are significantly different from observed historical trends.

One of the main motivations for creating Comet was to handle arbitrarily sized regions and polygon geometries.  Previous methods work only on single pixels and then drill down through the entire image stack to extract relevant statistics and visualize for just that one pixel.  This isn't great when you're interested in broader areas and aggregating stats over both space and time, and Comet helps to fill this gap.  Furthermore, although ARIMA is incorporated in other packages (i.e. SITS) to remove anomalies like clouds, Comet uses it in a new way, splitting data into a "before" and "after" event to automatically flag anomalies.  I think this is a novel application in a software context, but I don't claim it's the most novel thing about Comet in the text either.
