### 2019-09-12 CosmiQ/CometTS#16 @rmsare
Thanks for clarifying - very helpful.

Looks good to me, @jshermeyer @Kevin-Mattheus-Moerman .
