### 2019-09-13 #1047 @Kevin-Mattheus-Moerman
@jshermeyer looks good. We are ready for the next steps. Can you please:

- [ ] Can you deposit the release (version) of the reviewed software in Zenodo? Once you have archived the software there please report the DOI of the archived version here. Please ensure the authors and title on the Zenodo archive match those of the paper.
- [ ] Can you list the version number? (is it still v1.0 or has it changed)?

thanks
