+++ Requirements.txt
@@ -5,7 +5,7 @@ geopandas>=0.4.0
 rasterio>=1.0a12
 shapely=1.6.4.post2
 fiona=1.8.4 
-affine=2.2.2
+affine>=2.2.2
 rasterstats=0.13.0
 matplotlib=3.0.2
 seaborn=0.9.0
