+++ Requirements.txt
@@ -1,18 +1,18 @@
 #Requirements
-numpy=1.15.4
-pandas=0.24.1 
+numpy==1.15.4
+pandas==0.24.1 
 geopandas>=0.4.0 
 rasterio>=1.0a12
-shapely=1.6.4.post2
-fiona=1.8.4 
-affine>=2.2.2
-rasterstats=0.13.0
-matplotlib=3.0.2
-seaborn=0.9.0
-jupyter-client=5.2.4
-jupyter-core=4.4.0
-ipywidgets=7.4.2
-ipython=7.2.0 
-tqdm=4.31.1
-scipy=1.2.0
-gdal=2.4.0
+shapely==1.6.4.post2
+fiona==1.8.4 
+affine==2.2.2
+rasterstats==0.13.0
+matplotlib==3.0.2
+seaborn==0.9.0
+jupyter-client==5.2.4
+jupyter-core==4.4.0
+ipywidgets==7.4.2
+ipython==7.2.0 
+tqdm==4.31.1
+scipy==1.2.0
+gdal==2.4.0
