+++ setup.py
@@ -1,6 +1,6 @@
 from setuptools import setup, find_packages
 
-version = '1.1.4'
+version = '1.2.0'
 
 readme = ''
 
