### 2019-09-16 #1047 @jshermeyer
New version on Zenodo: https://zenodo.org/record/3418091
Version number is 1.2.0 which incorporates all the edits and comments from reviewer.  This version has also been pushed to pypi.

Thanks!
