### 2019-09-17 #1047 @kthyng
@jshermeyer:
Edits for paper:
- [ ] Spacing issues throughout. Make sure there are spaces next to parentheses, this is coming up around references mostly but other times too.
- [ ] ROI is used before being defined
- [ ] obligatory spelled wrong first page 2nd paragraph
- [ ] Figure 3 caption appears to be set up incorrectly.
- [ ] The line above Figure 3: What is this? Is it trying to link to its own DOI? That is out of date and shouldn't be in the references anyway, it is on the left side of the page on the first page of the paper.
