### 2019-09-23 ad36989
Update paper.md
