### 2019-09-23 6b42bf4
Update paper.bib
