+++ paper.bib
@@ -8,16 +8,6 @@
   url          = {https://doi.org/10.5281/zenodo.267110}
 }
 
-@misc{Shermeyer:2018,
-  author       = {Shermeyer, Jacob},
-  title        = {Comet Time Series (CometTS) Visualizer},
-  month        = jul,
-  year         = 2018,
-  note         = {https://github.com/CosmiQ/CometTS},
-  doi          = {10.5281/zenodo.1324419},
-  url          = {https://doi.org/10.5281/zenodo.1324419}
-}
-
 @article{Shermeyer_PR:2018,
 	doi = {10.1117/12.2325585},
 	url = {https://doi.org/10.1117/12.2325585},
@@ -42,7 +32,7 @@
 
 
 @misc{e-sensing/sits:2018,
-  author       = {Gilberto Camara; Rolf Simoes; Pedro R. Andrade; Victor Maus; alber sánchez; Luiz Fernando Ferreira Gomes de Assis; lorenalves; ywatacarvalho; Adeline Marinho Maciel; Lubia Vinhas; Gilberto Ribeiro de Queiroz},
+  author       = {Gilberto Camara; Rolf Simoes; Pedro R. Andrade; Victor Maus; alber sánchez; Luiz Fernando Ferreira Gomes de Assis; lorenalves; ywatacarvalho; Adeline Marinho Maciel; Lubia Vinhas and Gilberto Ribeiro de Queiroz},
   title        = {e-sensing/SITS - Satellite Image Time Series Analysis: Version 1.12.5},
   month        = dec,
   year         = 2018,
