+++ paper.md
@@ -21,15 +21,15 @@ bibliography: paper.bib
 
 Comet Time Series (``CometTS``) is a software package and tool that enables workflows for the analysis and visualization of a time series of satellite imagery for the data science and geographic science communities. The tool aims to enable population estimation research, land use change detection, or natural disaster monitoring using a range of data types.  ``CometTS`` calculates relevant statistical quantities (e.g., measures of central tendency and variation) and provides a visualization of their changes over time (Figure 1 and Figure 2). The package can help to serve as a rapid inspection workflow and analyze multiple regions of interest (ROI)'s in seconds.
 
-The package builds upon previously introduced tools like ``TSTools``[@Holden:2017], ``TIMESAT``[@Jonsson_Eklundh:2004], and ``Satellite Image Time Series Analysis (SITS)``[@e-sensing/sits:2018] to provide a partially automated approach for analyzing a time series of satellite images.  ``CometTS`` adds new functionality to the overhead time series analysis and visualization domain in a few ways:  Specifically, ``CometTS`` output includes user-specified statistics such as mean, median, and quartiles, across arbitrarily sized ROI's with polygon geometries. Furthermore, the option for anomaly detection is also included, and ``CometTS`` leverages an Auto-Regressive Integrated Moving Average (ARIMA) analysis to quantify trends and test if observations are significantly different from observed historical trends.  The obligatory functionality of cloud and snow masking over any ROI is also included. Finally, ``CometTS`` is coded in Python, and removes the GIS requirement that is typical for such analyses.
+The package builds upon previously introduced tools like ``TSTools`` [@Holden:2017], ``TIMESAT`` [@Jonsson_Eklundh:2004], and ``Satellite Image Time Series Analysis (SITS)`` [@e-sensing/sits:2018] to provide a partially automated approach for analyzing a time series of satellite images.  ``CometTS`` adds new functionality to the overhead time series analysis and visualization domain in a few ways:  Specifically, ``CometTS`` output includes user-specified statistics such as mean, median, and quartiles, across arbitrarily sized ROI's with polygon geometries. Furthermore, the option for anomaly detection is also included, and ``CometTS`` leverages an Auto-Regressive Integrated Moving Average (ARIMA) analysis to quantify trends and test if observations are significantly different from observed historical trends.  The obligatory functionality of cloud and snow masking over any ROI is also included. Finally, ``CometTS`` is coded in Python, and removes the GIS requirement that is typical for such analyses.
 
 ![Figure 1](https://raw.githubusercontent.com/CosmiQ/CometTS/master/ExamplePlots/Workflow.png)
 **Fig. 1:** A four step process to analyze and visualize trends in a time series of overhead imagery.
 
 ![Figure 2](https://raw.githubusercontent.com/CosmiQ/CometTS/master/ExamplePlots/Niamey.png)
-**Fig. 2:** An example output from CometTS showing variations in nighttime brightness in Niamey, Niger.  These seasonal trends can be correlated to season migration patterns in the city.
+**Fig. 2:** An example output from ``CometTS`` showing variations in nighttime brightness in Niamey, Niger.  These seasonal trends can be correlated to season migration patterns in the city.
 
[…]
