### 2019-09-23 #1047 @jshermeyer
@kthyng I think I have fixed all of the issues you flagged (among a few others).  Let me know if there's anything else you notice.  Thanks!
