### 2019-09-26 #1047 @kthyng
> Another edit, @jshermeyer: your citation e-sensing/sits:2018 is not formatted correctly. In particular at least the authors in the author list need to be separated by "and" as in your other entries. There may be other problems too, but you can see in your reference list and when you cite it inline that it is not working correctly.

@jshermeyer This doesn't appear to have changed yet.
