### 2019-09-26 #1047 @kthyng
[Here](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) under "example paper.bib", the first example paper shows how multiple authors (each author) should be separated by "and" instead of a semi-colon. You can check to make sure it worked properly and there aren't other errors by how it looks when cited (shouldn't take up a whole line in the paper) and how it looks in the references (some of the authors are being abbreviated to a single letter I think?).
