+++ paper.bib
@@ -32,7 +32,7 @@
 
 
 @misc{e-sensing/sits:2018,
-  author       = {Gilberto Camara; Rolf Simoes; Pedro R. Andrade; Victor Maus; alber sánchez; Luiz Fernando Ferreira Gomes de Assis; lorenalves; ywatacarvalho; Adeline Marinho Maciel; Lubia Vinhas and Gilberto Ribeiro de Queiroz},
+  author       = {Gilberto Camara and Rolf Simoes and Pedro R. Andrade and Victor Maus and alber sánchez and Luiz Fernando Ferreira Gomes de Assis and lorenalves and ywatacarvalho and Adeline Marinho Maciel and Lubia Vinhas and Gilberto Ribeiro de Queiroz},
   title        = {e-sensing/SITS - Satellite Image Time Series Analysis: Version 1.12.5},
   month        = dec,
   year         = 2018,
