### 2019-10-01 #1047 @jshermeyer
Thanks all  @zhampel, @rmsare, @jjmcnelis, @Kevin-Mattheus-Moerman, @kthyng !

This is presently giving me a 404:
https://joss.theoj.org/papers/10.21105/joss.01047

Assuming it will auto-update given a bit of time?
