### 2019-10-01 #1047 @jshermeyer
Working on another browser, all good! I must just have to clear my cookies or something.
